### 2025-04-11 #8042 @editorialbot
[PRE REVIEW]: maze-dataset: Maze Generation with Algorithmic Variety and Representational Flexibility
**Submitting author:** @mivanit (<a href="http://orcid.org/0000-0002-4213-4993">Michael Ivanitskiy</a>)
**Repository:** https://github.com/understanding-search/maze-dataset
**Branch with paper.md** (empty if default branch): joss-paper
**Version:** v1.3.2
