### 2025-04-26 385af3a
move figure code to external files
