+++ docs/paper/paper.tex
@@ -489,15 +489,7 @@ various visualizations. As well as providing a simple interface for
 generating, storing, and loading these datasets, \texttt{maze-dataset}
 is extensively tested, type hinted, benchmarked, and documented.
 
-\begin{figure} 
-  \begin{minipage}{5in}
-    \input{diagram/diagram.tikz} 
-  \end{minipage}
-  \caption{
-    Usage of maze-dataset. We create a \texttt{MazeDataset} from a \texttt{MazeDatasetConfig}. This contains \texttt{SolvedMaze} objects which can be converted to and from a variety of formats. Code in the image contains clickable links to \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html}{documentation}. A variety of generated examples can be viewed \href{https://understanding-search.github.io/maze-dataset/examples/maze_examples.html}{here}.
-  }
-  \label{fig:diagram}
-\end{figure}
+\input{figures/tex/fig1_diagram.tex}
 
 \hypertarget{statement-of-need}{%
 \section{Statement of Need}\label{statement-of-need}}
@@ -708,36 +700,7 @@ objects, which have path information and a tensor optimized for storing
 sub-graphs of a lattice. These objects can be converted to and from
 several formats to maximize their utility in different contexts.
 
-\begin{figure}[H]
-  \centering
-  \begin{tabular}{p{1.5in} p{1.5in} p{1.5in}} 
-    \hline \\[.5em]
-    % algorithms
-    \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}
-    & \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_pixels}{\texttt{as\_pixels()}}
-    & \href{https://understanding-search.github.io/maze-dataset/maze_dataset/plotting.html\#MazePlot}{\texttt{MazePlot()}} \\[.5em]
-    % descriptions
-      Simple text format for displaying mazes, useful for debugging in a terminal environment.
-      & \texttt{numpy} array of \texttt{dtype=uint8} and shape \texttt{(height, width, 3)}. The last dimension is RGB color.
-      & feature-rich plotting utility with support for multiple paths, heatmaps over positions, and more. \\[1em]
-    \hline \\
-    % examples
-      \multicolumn{1}{c}{\begin{minipage}[b]{1.6in}
-        \setlength{\baselineskip}{0.9em}
-        \input{figures/outputs-ascii-colored.tex} 
-      \end{minipage}}
-      & \multicolumn{1}{c}{
-        \includegraphics[width=0.25\textwidth]{figures/outputs-pixels.pdf}
-      }
-      & \multicolumn{1}{c}{
-        \includegraphics[width=0.27\textwidth, trim={0 0.8cm -.3cm, -.5cm}, clip]{figures/outputs-mazeplot.pdf}
-      } \\[1em]
-    
-    \hline \\
-  \end{tabular}
-  \caption{Various output formats. Top row (left to right): ASCII diagram, rasterized pixel grid, and advanced display tool.}
-  \label{fig:output-fmts}
-\end{figure}
+\input{figures/tex/fig2_formats.tex}
 
 In previous work, maze tasks have been used with Recurrent Convolutional
 Neural Network (RCNN) derived architectures
@@ -804,21 +767,7 @@ All output sequences consist of four token regions representing
 different features of the maze; an example output sequence is shown in
 \autoref{fig:token-regions}.
 
-\begin{figure} 
-  \centering
-  \begin{minipage}{5in}
-    \footnotesize
-    \input{figures/outputs-tokens-colored.tex}
-  \end{minipage}
[…]
