+++ docs/paper/paper.md
@@ -67,19 +67,7 @@ header-includes: |
 
 Solving mazes is a classic problem in computer science and artificial intelligence, and humans have been constructing mazes for thousands of years. Although finding the shortest path through a maze is a solved problem, this very fact makes it an excellent testbed for studying how machine learning algorithms solve problems and represent spatial information. We introduce `maze-dataset`, a user-friendly Python library for generating, processing, and visualizing datasets of mazes. This library supports a variety of maze generation algorithms providing mazes with or without loops, mazes that are connected or not, and many other variations. These generation algorithms can be configured with various parameters, and the resulting mazes can be filtered to satisfy desired properties. Also provided are tools for converting mazes to and from various formats suitable for a variety of neural network architectures, such as rasterized images, tokenized text sequences, and various visualizations. As well as providing a simple interface for generating, storing, and loading these datasets, `maze-dataset` is extensively tested, type hinted, benchmarked, and documented.
 
-\begin{figure} 
-  \begin{minipage}{5in}
-    \input{diagram/diagram.tikz} 
-  \end{minipage}
-  \caption{
-    Usage of maze-dataset. We create a \texttt{MazeDataset} from a \texttt{MazeDatasetConfig}. This contains \texttt{SolvedMaze} objects which can be converted to and from a variety of formats. Code in the image contains clickable links to \docslink{maze_dataset.html}{documentation}. A variety of generated examples can be viewed \docslink{examples/maze_examples.html}{here}.
-  }
-  \label{fig:diagram}
-\end{figure}
-
-<!-- ```{=html}
-<img src="diagram/diagram.svg"/>
-``` -->
+\input{figures/tex/fig1_diagram.tex}
 
 # Statement of Need
 
@@ -154,36 +142,7 @@ For use cases where mazes of different sizes, generation algorithms, or other pa
 
 Internally, mazes are [`SolvedMaze`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#SolvedMaze) objects, which have path information and a tensor optimized for storing sub-graphs of a lattice. These objects can be converted to and from several formats to maximize their utility in different contexts.
 
-\begin{figure}[H]
-  \centering
-  \begin{tabular}{p{1.5in} p{1.5in} p{1.5in}} 
-    \hline \\[.5em]
-    % algorithms
-    \docslink{maze_dataset.html\#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}
-    & \docslink{maze_dataset.html\#LatticeMaze.as_pixels}{\texttt{as\_pixels()}}
-    & \docslink{maze_dataset/plotting.html\#MazePlot}{\texttt{MazePlot()}} \\[.5em]
-    % descriptions
-      Simple text format for displaying mazes, useful for debugging in a terminal environment.
-      & \texttt{numpy} array of \texttt{dtype=uint8} and shape \texttt{(height, width, 3)}. The last dimension is RGB color.
-      & feature-rich plotting utility with support for multiple paths, heatmaps over positions, and more. \\[1em]
-    \hline \\
-    % examples
-      \multicolumn{1}{c}{\begin{minipage}[b]{1.6in}
-        \setlength{\baselineskip}{0.9em}
-        \input{figures/outputs-ascii-colored.tex} 
-      \end{minipage}}
-      & \multicolumn{1}{c}{
-        \includegraphics[width=0.25\textwidth]{figures/outputs-pixels.pdf}
-      }
-      & \multicolumn{1}{c}{
-        \includegraphics[width=0.27\textwidth, trim={0 0.8cm -.3cm, -.5cm}, clip]{figures/outputs-mazeplot.pdf}
[…]
