+++ docs/paper/figures/tex/fig1_diagram.tex
@@ -1,0 +1,9 @@
+\begin{figure} 
+	\begin{minipage}{5in}
+	  \input{diagram/diagram.tikz} 
+	\end{minipage}
+	\caption{
+	  Usage of maze-dataset. We create a \texttt{MazeDataset} from a \texttt{MazeDatasetConfig}. This contains \texttt{SolvedMaze} objects which can be converted to and from a variety of formats. Code in the image contains clickable links to \docslink{maze_dataset.html}{documentation}. A variety of generated examples can be viewed \docslink{examples/maze_examples.html}{here}.
+	}
+	\label{fig:diagram}
+  \end{figure}

[…]
