+++ docs/paper/paper.md
@@ -124,7 +124,7 @@ dataset: MazeDataset = MazeDataset.from_config(
   cfg, # pass the config
   ..., # other options for disk loading, parallelization, etc.
 )
-```
+``` 
 
 When initializing a dataset, options which do not affect the mazes themselves can be specified through the [`from_config()`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#MazeDataset.from_config) factory method as necessary. These options allow for saving/loading existing datasets instead of re-generating, parallelization options for generation, and more. Available maze generation algorithms are static methods of the [`LatticeMazeGenerators`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#LatticeMazeGenerators) namespace class and include generation algorithms based on randomized depth-first search, Wilson's algorithm [@wilson], percolation [@percolation; @percolation-clustersize], Kruskal's algorithm [@kruskal1956shortest], and others.
 
@@ -161,13 +161,13 @@ All output sequences consist of four token regions representing different featur
 
 \input{figures/tex/fig4_tokenfmt.tex}
 
-Each [`MazeTokenizerModular`](https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html#MazeTokenizerModular) is constructed from a set of several [`_TokenizerElement`](https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html#_TokenizerElement) objects, each of which specifies how different token regions or other elements of the stringification are produced.
+Each [`MazeTokenizerModular`](https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html#MazeTokenizerModular) is constructed from a set of several \docslinkcode{maze_dataset/tokenization.html#_TokenizerElement}{\_TokenizerElement} objects, each of which specifies how different token regions or other elements of the stringification are produced.
 
 \input{figures/tex/fig5_mmt.tex}
 
-The tokenizer architecture is purposefully designed such that adding and testing a wide variety of new tokenization algorithms is fast and minimizes disturbances to functioning code. This is enabled by the modular architecture and the automatic inclusion of any new tokenizers in integration tests. To create a new variety of tokenizer, developers forking the library may simply create their own [`_TokenizerElement`](https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html#_TokenizerElement) subclass and implement the abstract methods. If the behavior change is sufficiently small, simply adding a parameter to an existing [`_TokenizerElement`](https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html#_TokenizerElement) subclass and updating its implementation will suffice.
+The tokenizer architecture is purposefully designed such that adding and testing a wide variety of new tokenization algorithms is fast and minimizes disturbances to functioning code. This is enabled by the modular architecture and the automatic inclusion of any new tokenizers in integration tests. To create a new variety of tokenizer, developers forking the library may simply create their own \docslinkcode{maze_dataset/tokenization.html#_TokenizerElement}{\_TokenizerElement} subclass and implement the abstract methods. If the behavior change is sufficiently small, simply adding a parameter to an existing \docslinkcode{maze_dataset/tokenization.html#_TokenizerElement}{\_TokenizerElement} subclass and updating its implementation will suffice.
 
[…]
