### 2025-04-26 537ddba
cut down impl and sre sections, better links
