+++ docs/paper/paper.md
@@ -60,6 +60,7 @@ header-includes: |
   \providecommand{\XeTeXLinkBox}[1]{#1}
   \newcommand{\docslink}[2]{\href{https://understanding-search.github.io/maze-dataset/#1}{#2}}
   \newcommand{\docslinkcode}[2]{\href{https://understanding-search.github.io/maze-dataset/#1}{\texttt{#2}}}
+  \newcommand{\docslinkcodemain}[1]{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\##1}{\texttt{#1}}}
   \newcommand{\secref}[1]{\hyperref[#1]{section: \textit{\nameref{#1}}}}
 ---
 
@@ -102,7 +103,7 @@ We direct readers to our [examples](https://understanding-search.github.io/maze-
 
 Our package can be installed from [PyPi](https://pypi.org/project/maze-dataset/) via `pip install maze-dataset`, or directly from the [git repository](https://github.com/understanding-search/maze-dataset) [@maze-dataset-github].
 
-To create a dataset, we first create a [`MazeDatasetConfig`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#MazeDatasetConfig) configuration object, which specifies the seed, number, and size of mazes, as well as the generation algorithm and its corresponding parameters. This object is passed to a [`MazeDataset`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#MazeDataset) class to create a dataset. Crucially, this [`MazeDataset`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#MazeDataset) mimics the interface of a PyTorch [@pytorch] [`Dataset`](https://pytorch.org/docs/stable/data.html), and can thus be easily incorporated into existing data pre-processing and training pipelines, e.g., through the use of a `DataLoader` class.
+To create a dataset, we first create a \docslinkcodemain{MazeDatasetConfig} configuration object, which specifies the seed, number, and size of mazes, as well as the generation algorithm and its corresponding parameters. This object is passed to a \docslinkcodemain{MazeDataset} class to create a dataset. Crucially, this \docslinkcodemain{MazeDataset} mimics the interface of a PyTorch [@pytorch] [`Dataset`](https://pytorch.org/docs/stable/data.html), and can thus be easily incorporated into existing data pre-processing and training pipelines, e.g., through the use of a `DataLoader` class.
 
 ```python
 from maze_dataset import (
@@ -134,7 +135,7 @@ Furthermore, a dataset of mazes can be filtered to satisfy certain properties. C
 dataset_filtered: MazeDataset = dataset.filter_by.path_length(min_length=3)
 ```
 
-All implemented maze generation algorithms are stochastic by nature. For reproducibility, the `seed` parameter of [`MazeDatasetConfig`](https://understanding-search.github.io/maze-dataset/maze_dataset.html#MazeDatasetConfig) may be set. In practice, using provided deduplication filters, we find that exact duplicate mazes are generated very infrequently, even when generating very large datasets.
+All implemented maze generation algorithms are stochastic by nature. For reproducibility, the `seed` parameter of \docslinkcodemain{MazeDatasetConfig} may be set. In practice, using provided deduplication filters, we find that exact duplicate mazes are generated very infrequently, even when generating very large datasets.
 
 For use cases where mazes of different sizes, generation algorithms, or other parameter variations are required, we provide the [`MazeDatasetCollection`](https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/collected_dataset.html#MazeDatasetCollection) class, which allows for creating a single iterable dataset from multiple independent configurations.
 
@@ -175,57 +176,23 @@ We provide approximate benchmarks for relative generation time across various al
 
 \input{figures/tex/tab1_benchmarks.tex}
[…]
