+++ docs/paper/paper.tex
@@ -363,6 +363,7 @@ Software}, \emph{¿VOL?}(¿ISSUE?), ¿PAGE? \url{https://doi.org/N/A}}.}}
 \providecommand{\XeTeXLinkBox}[1]{#1}
 \newcommand{\docslink}[2]{\href{https://understanding-search.github.io/maze-dataset/#1}{#2}}
 \newcommand{\docslinkcode}[2]{\href{https://understanding-search.github.io/maze-dataset/#1}{\texttt{#2}}}
+\newcommand{\docslinkcodemain}[1]{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\##1}{\texttt{#1}}}
 \newcommand{\secref}[1]{\hyperref[#1]{section: \textit{\nameref{#1}}}}
 \ifLuaTeX
   \usepackage{selnolig}  % disable illegal ligatures
@@ -814,21 +815,7 @@ without parallelism.
 
 \input{figures/tex/tab1_benchmarks.tex}
 
-\begin{figure}
-\hypertarget{fig:benchmarks}{%
-\centering
-\includegraphics[width=0.9\textwidth,height=\textheight]{figures/benchmarks/gridsize-vs-gentime.pdf}
-\caption{Plot of maze generation time. Generation time scales
-exponentially with maze size for all algorithms. Generation time per
-maze does not depend on the number of mazes being generated, and there
-is minimal overhead to initializing the generation process for a small
-dataset. Wilson's algorithm is notably less efficient than others and
-has high variance. Note that values are averaged across all parameter
-sets for that algorithm. More information can be found on the
-\href{https://understanding-search.github.io/maze-dataset/benchmarks/}{benchmarks
-page}.}\label{fig:benchmarks}
-}
-\end{figure}
+\input{figures/tex/fig6_benchmarks.tex}
 
 \hypertarget{sec:success-rate-estimation}{%
 \subsection{Success Rate Estimation}\label{sec:success-rate-estimation}}
@@ -851,87 +838,20 @@ configuration using a symbolic regression model trained with PySR
 (\protect\hyperlink{ref-pysr}{Cranmer, 2023}). More details on this can
 be found in
 \href{https://understanding-search.github.io/maze-dataset/notebooks/estimate_dataset_fractions.html}{\texttt{estimate\_dataset\_fractions.ipynb}}.
-Using the estimation algorithm simply requires the user to call
-\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#MazeDatasetConfig.success_fraction_compensate}{\texttt{cfg\_new:\ MazeDatasetConfig\ =\ cfg.success\_fraction\_compensate()}},
-providing their initial \texttt{cfg} and then using the returned
-\texttt{cfg\_new} in its place.
 
-\hypertarget{success-rate-estimation-algorithm}{%
-\subsubsection{Success Rate Estimation
-Algorithm}\label{success-rate-estimation-algorithm}}
-
-The base function learned by symbolic regression provides limited
-insight and may be subject to change. It is defined as
-\href{https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/success_predict_math.html\#cfg_success_predict_fn}{\texttt{cfg\_success\_predict\_fn}},
-and takes a 5 dimensional float vector created by
-\texttt{MazeDatasetConfig.\_to\_ps\_array()} which represents the
-{[}percolation value, grid size, endpoint deadend configuration,
-endpoint uniqueness, categorical generation function index{]}.
-
-However, the outputs of this function are not directly usable due to
-minor divergences at the endpoints with respect to the percolation
[…]
