### 2025-04-26 0e30365
cut down tokenization section
