+++ docs/paper/paper.tex
@@ -316,15 +316,6 @@ Software}, \emph{¿VOL?}(¿ISSUE?), ¿PAGE? \url{https://doi.org/N/A}}.}}
 \newcommand{\VerbatimStringTok}[1]{\textcolor[rgb]{0.25,0.44,0.63}{#1}}
 \newcommand{\WarningTok}[1]{\textcolor[rgb]{0.38,0.63,0.69}{\textbf{\textit{#1}}}}
 
-\usepackage{graphicx,grffile}
-\makeatletter
-\def\maxwidth{\ifdim\Gin@nat@width>\linewidth\linewidth\else\Gin@nat@width\fi}
-\def\maxheight{\ifdim\Gin@nat@height>\textheight\textheight\else\Gin@nat@height\fi}
-\makeatother
-% Scale images if necessary, so that they will not overflow the page
-% margins by default, and it is still possible to overwrite the defaults
-% using explicit options in \includegraphics[width, height, ...]{}
-\setkeys{Gin}{width=\maxwidth,height=\maxheight,keepaspectratio}
 \IfFileExists{parskip.sty}{%
 \usepackage{parskip}
 }{% else
@@ -567,20 +558,19 @@ with these datasets.
   The text format provided by
   \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#MazeDataset.as_tokens}{\texttt{SolvedMaze(...).as\_tokens()}}
   is similar to that of (\protect\hyperlink{ref-eval-LLM-graphs}{Liu \&
-  Wu, 2023}), but provides over 5.8 million unique formats for
-  converting mazes to a text stream, detailed in
+  Wu, 2023}) but with many more options, detailed
   \hyperref[sec:tokenized-output-formats]{section: \textit{\nameref{sec:tokenized-output-formats}}}.
 \item
-  For rigorous investigations of the response of a model to various
-  distributional shifts, preserving metadata about the generation
-  algorithm with the dataset itself is essential. To this end, our
-  package efficiently stores the dataset along with its metadata in a
-  single human-readable file (\protect\hyperlink{ref-zanj}{M.
+  Preserving metadata about the generation algorithm with the dataset
+  itself is essential for studying the effects of distributional shifts.
+  Our package efficiently stores the dataset along with its metadata in
+  a single human-readable file (\protect\hyperlink{ref-zanj}{M.
   Ivanitskiy, n.d.}). As far as we are aware, no existing packages do
   this reliably.
 \item
-  Storing mazes as images is not only difficult to work with, but also
-  inefficient. We use a highly efficient method detailed in
+  Storing mazes as images or adjacency matrices is not only difficult to
+  work with, but also inefficient. We use a highly efficient method
+  detailed in
   \hyperref[sec:implementation]{section: \textit{\nameref{sec:implementation}}}.
 \item
   Our package is easily installable with source code freely available.
@@ -714,16 +704,7 @@ Bansal, et al., 2021}) and provide the
 class which returns rasterized pairs of (input, target) mazes as shown
 in \autoref{fig:e2h-raster} below.
 
-\begin{figure}
-\hypertarget{fig:e2h-raster}{%
-\centering
-\includegraphics[width=0.3\textwidth,height=\textheight]{figures/maze-raster-input-target.pdf}
-\caption{Input is the rasterized maze without the path marked (left),
-and provide as a target the maze with all but the correct path removed
-(right). Configuration options exist to adjust whether endpoints are
-included and if empty cells should be filled in.}\label{fig:e2h-raster}
-}
-\end{figure}
+\input{figures/tex/fig3_raster.tex}
 
 \hypertarget{sec:tokenized-output-formats}{%
[…]
