### 2025-04-26 #8042 @ekourlit
I'm happy to take over this paper @crvernon!

Greetings @mivanit, I will be the editor taking care of your submission. After having an initial look at the repository and the paper, your submission fulfills the minimal requirements to proceed with the review. However, I would need to ask you to reduce the length of your paper to 1000 lines. You can remove details on the architecture of the software, as software documentation such as API functionality should not be in the paper and instead should be outlined in the software documentation pages. Additionally, as you see above, there are problems with many DOIs, could you please address?

Do you already have in mind a couple of individuals that might be willing to review this work for JOSS ? In general, we would appreciate people with relevant domain knowledge. Please just have potential [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) in mind. You can consult the automatic suggestions https://github.com/openjournals/joss-reviews/issues/7680#issuecomment-2596321673, or will have a look at the [JOSS reviewer database](https://reviewers.joss.theoj.org/).
