### 2025-04-27 #8042 @amcandio
Hi! I’d be happy to volunteer as a reviewer for this submission. I’m a tenured software engineer with experience in scientific Python software, graph algorithms, and open-source contributions through projects like NetworkX.
