### 2025-05-28 ee796b5
fix some figures, cut some sections
