+++ docs/paper/paper.md
@@ -40,28 +40,7 @@ affiliations:
     index: 4
 date: 9 April 2025
 bibliography: refs.bib
-header-includes: |
-  \usepackage{graphicx}
-  \usepackage{tikz}
-  \usetikzlibrary{calc}
-  \tikzset{ % Define a TikZ style for an external hyperlink node.
-    hyperlink node url/.style={
-      alias=sourcenode,
-      append after command={
-        let \p1 = (sourcenode.north west),
-            \p2 = (sourcenode.south east),
-            \n1 = {\x2-\x1},
-            \n2 = {\y1-\y2} in
-        node[inner sep=0pt, outer sep=0pt, anchor=north west, at=(\p1)]
-            {\href{#1}{\XeTeXLinkBox{\phantom{\rule{\n1}{\n2}}}}}
-      }
-    }
-  }
-  \providecommand{\XeTeXLinkBox}[1]{#1}
-  \newcommand{\docslink}[2]{\href{https://understanding-search.github.io/maze-dataset/#1}{#2}}
-  \newcommand{\docslinkcode}[2]{\href{https://understanding-search.github.io/maze-dataset/#1}{\texttt{#2}}}
-  \newcommand{\docslinkcodemain}[1]{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\##1}{\texttt{#1}}}
-  \newcommand{\secref}[1]{\hyperref[#1]{section: \textit{\nameref{#1}}}}
+header-includes: \input{preamble.tex}
 ---
 
 # Summary
@@ -97,47 +76,10 @@ A multitude of public and open-source software packages exist for generating maz
 
 # Features
 
-We direct readers to our [examples](https://understanding-search.github.io/maze-dataset/examples/maze_examples.html), [docs](https://understanding-search.github.io/maze-dataset/maze_dataset.html), and [notebooks](https://understanding-search.github.io/maze-dataset/notebooks/) for more information.
+We direct readers to our [examples](https://understanding-search.github.io/maze-dataset/examples/maze_examples.html), [docs](https://understanding-search.github.io/maze-dataset/maze_dataset.html), and [notebooks](https://understanding-search.github.io/maze-dataset/notebooks/) for more information. Our package can be installed from [PyPi](https://pypi.org/project/maze-dataset/) via `pip install maze-dataset`, or directly from the [git repository](https://github.com/understanding-search/maze-dataset) [@maze-dataset-github].
 
-## Generation and Basic Usage {#generation}
+Datasets of mazes are created from a \docslinkcodemain{MazeDatasetConfig} configuration object, which allows specifying the number of mazes, their size, the generation algorithm, and various parameters for the generation algorithm. Datasets can also be filtered after generation to satisfy certain properties. Custom filters can be specified, and some filters are included in [`MazeDatasetFilters`](https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/filters.html#MazeDatasetFilters)
 
-Our package can be installed from [PyPi](https://pypi.org/project/maze-dataset/) via `pip install maze-dataset`, or directly from the [git repository](https://github.com/understanding-search/maze-dataset) [@maze-dataset-github].
-
[…]
