+++ docs/paper/preamble.tex
@@ -1,4 +1,5 @@
 \usepackage{graphicx}
+\usepackage{array}
 \usepackage{tikz}
 \usetikzlibrary{calc}
 \tikzset{ % Define a TikZ style for an external hyperlink node.

[…]
