### 2025-05-28 bc4f76e
ascii maze finally fixed
