+++ docs/paper/figures/outputs-ascii-colored.tex
@@ -1,12 +1,14 @@
-{\ttfamily
-\setlength{\tabcolsep}{0.5em}
-\setlength{\arraystretch}{0.8}
+{
+% \footnotesize
+\ttfamily
+\setlength{\tabcolsep}{0.3em}
+\renewcommand{\arraystretch}{0.75}
 \begin{tabular}{ccccccccccc}
 \# & \# & \# & \# & \# & \# & \# & \# & \# & \# & \# \\
 \# &   &   &   &   & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \# &  & \# \\
 \# &   & \# & \# & \# & \textcolor{blue}{X} & \# & \textcolor{blue}{X} & \# &   & \# \\
 \# &   &   &   & \# & \textcolor{blue}{X} & \# & \textcolor{green}{S} &   &   & \# \\
-\# & \# & \# & \# & \# & \textcolor{blue}{X} & \# & \# & \# & \# & \# \\
+\# & \# & \# & \# & \# & \textcolor{b			lue}{X} & \# & \# & \# & \# & \# \\
 \# & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \# & \textcolor{red}{E} & \textcolor{blue}{X} & \textcolor{blue}{X} & \# \\
 \# & \textcolor{blue}{X} & \# & \# & \# &   & \# & \# & \# & \textcolor{blue}{X} & \# \\
 \# & \textcolor{blue}{X} & \# &   &   &   &   &   & \# & \textcolor{blue}{X} & \# \\
