+++ docs/paper/figures/tex/fig2_formats.tex
@@ -12,14 +12,18 @@
 		& feature-rich plotting utility with support for multiple paths, heatmaps over positions, and more. \\[1em]
 	  \hline \\
 	  % examples
-		\multicolumn{1}{c}{\begin{minipage}[b]{1.6in}
-		  \setlength{\baselineskip}{0.9em}
-		  \input{figures/outputs-ascii-colored.tex} 
-		\end{minipage}}
+		\multicolumn{1}{l}{
+		  \hspace{-1.5em}\raisebox{0.85\height}{
+			  \input{figures/outputs-ascii-colored.tex}
+		  }
+		}
+		% 	\begin{minipage}[b]{1.6in}
+		%   \setlength{\baselineskip}{0.9em}
+		% 	\end{minipage}
 		& \multicolumn{1}{c}{
 		  \includegraphics[width=0.25\textwidth]{figures/outputs-pixels.pdf}
 		}
-		& \multicolumn{1}{c}{
+		& \multicolumn{1}{r}{
 		  \includegraphics[width=0.27\textwidth, trim={0 0.8cm -.3cm, -.5cm}, clip]{figures/outputs-mazeplot.pdf}
 		} \\[1em]
 	  
