+++ docs/paper/paper.tex
@@ -583,16 +583,7 @@ reasoning steps (\protect\hyperlink{ref-pfau2024dotbydot}{Pfau et al.,
 2024}; \protect\hyperlink{ref-spies2024causalworldmodels}{Spies et al.,
 2024}). To facilitate systematic investigation of the effects of
 different representations of data on text model performance, we provide
-a variety of text output formats. By passing an instance of
-\href{https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html\#MazeTokenizerModular}{\texttt{MazeTokenizerModular}}
-to
-\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#MazeDataset.as_tokens}{\texttt{as\_tokens(...)}},
-a maze can be converted to a text sequence. The
-\href{https://understanding-search.github.io/maze-dataset/maze_dataset/tokenization.html\#MazeTokenizerModular}{\texttt{MazeTokenizerModular}}
-class contains a rich set of options with class contains 19 discrete
-parameters, resulting in over 5.8 million unique possible tokenizers.
-All output sequences consist of four token regions representing
-different features of the maze; an example output sequence is shown in
+a variety of text output formats, with an example given in
 \autoref{fig:token-regions}.
 
 \input{figures/tex/fig4_tokenfmt.tex}
