+++ docs/paper/figures/tex/tab1_benchmarks.tex
@@ -35,6 +35,6 @@
 			&                        &   11.1   &    6.5   &   32.9   &  302.7   \\
 	\hline
 \end{tabular}
-\caption{Generation times for various algorithms and maze sizes. More information can be found on the \docslink{benchmarks}{benchmarks page}.}
+\caption{Generation times in milliseconds for various algorithms and maze sizes. More information can be found on the \docslink{benchmarks}{benchmarks page}.}
 \label{tab:benchmarks}
 \end{table}
