### 2025-05-28 3156e22
add some DOIs
