+++ docs/paper/refs.bib
@@ -1,9 +1,10 @@
 @article{dijkstra-prim,
-    title = {A Note on Two Problems in Connexion with Graphs:({{Numerische Mathematik}}, 1 (1959), p 269-271)},
+    title = { :({{Numerische Mathematik}}, 1 (1959), p 269-271)},
     shorttitle = {A Note on Two Problems in Connexion with Graphs},
     author = {Dijkstra, Edsger Wybe},
     year = {1959},
     publisher = {{Stichting Mathematisch Centrum}},
+    doi = {10.1007/bf01386390},
 }
 
 @inproceedings{fung2022jfb,
@@ -13,28 +14,32 @@
   volume={36},
   number={6},
   pages={6648--6656},
-  year={2022}
+  year={2022},
+  doi = {10.1609/aaai.v36i6.20619},
 }
 
 @article{eval-LLM-graphs,
     title={Evaluating Large Language Models on Graphs: Performance Insights and Comparative Analysis},
     author={Liu, Chang and Wu, Bo},
     journal={arXiv preprint arXiv:2308.11224},
-    year={2023}
+    year={2023},
+    doi = {10.48550/arXiv.2308.11224},
 }
 
 @article{zhang2025tscend,
   title={T-SCEND: Test-time Scalable MCTS-enhanced Diffusion Model},
   author={Zhang, Tao and Pan, Jia-Shu and Feng, Ruiqi and Wu, Tailin},
   journal={arXiv preprint arXiv:2502.01989},
-  year={2025}
+  year={2025},
+  doi = {10.48550/arXiv.2502.01989},
 }
 
 @article{dao2025alphamaze,
   title={AlphaMaze: Enhancing Large Language Models' Spatial Intelligence via GRPO},
   author={Dao, Alan and Vu, Dinh Bach},
   journal={arXiv preprint arXiv:2502.14669},
-  year={2025}
+  year={2025},
+  doi = {10.48550/arXiv.2502.14669},
 }
 
 
@@ -72,7 +77,8 @@
   title={Interpretable machine learning for science with PySR and SymbolicRegression. jl},
   author={Cranmer, Miles},
   journal={arXiv preprint arXiv:2305.01582},
-  year={2023}
+  year={2023},
+  doi = {10.48550/arXiv.2305.01582},
 }
 
 
@@ -127,13 +133,15 @@
     url = {http://arxiv.org/abs/1712.04651},
     urldate = {2023-09-02},
     abstract = {Percolation models describe the inside of a porous material. The theory emerged timidly in the middle of the twentieth century before becoming one of the major objects of interest in probability and mathematical physics. The golden age of percolation is probably the eighties, during which most of the major results were obtained for the most classical of these models, named Bernoulli percolation, but it is really the two following decades which put percolation theory at the crossroad of several domains of mathematics. In this short (and very partial) review, we propose to describe briefly some of the recent progress as well as some famous challenges remaining in the field. This review is not intended to probabilists (and a fortiori not to specialists in percolation theory): the target audience is mathematicians of all kinds.},
+    doi = {10.48550/arXiv.1712.04651},
 }
 
 @article{mckenzie2023faster,
     title={Faster predict-and-optimize with three-operator splitting},
     author={McKenzie, Daniel and Fung, Samy Wu and Heaton, Howard},
     journal={arXiv preprint arXiv:2301.13395},
-    year={2023}
+    year={2023},
[…]
