### 2025-05-29 2951a2c
add some DOIs
