+++ docs/paper/paper.tex
@@ -686,20 +686,23 @@ Ayaz, H., Allen, S. L., Platek, S. M., \& Onaral, B. (2008). Maze suite
 1.0: A complete set of tools to prepare, present, and analyze
 navigational and spatial cognitive neuroscience experiments.
 \emph{Behavior Research Methods}, \emph{40}, 353--359.
+\url{https://doi.org/10.3758/brm.40.1.353}
 
 \leavevmode\vadjust pre{\hypertarget{ref-chen2024iaimperative}{}}%
 Chen, X., Yang, F., \& Wang, C. (2024). {iA*}: Imperative learning-based
 {A*} search for pathfinding. \emph{arXiv Preprint arXiv:2403.15870}.
+\url{https://doi.org/10.48550/arXiv.2403.15870}
 
 \leavevmode\vadjust pre{\hypertarget{ref-cobbe2019procgen}{}}%
 Cobbe, K., Hesse, C., Hilton, J., \& Schulman, J. (2019). Leveraging
 procedural generation to benchmark reinforcement learning. \emph{arXiv
 Preprint arXiv:1912.01588}.
+\url{https://doi.org/10.48550/arXiv.1912.01588}
 
 \leavevmode\vadjust pre{\hypertarget{ref-dao2025alphamaze}{}}%
 Dao, A., \& Vu, D. B. (2025). AlphaMaze: Enhancing large language
 models' spatial intelligence via GRPO. \emph{arXiv Preprint
-arXiv:2502.14669}.
+arXiv:2502.14669}. \url{https://doi.org/10.48550/arXiv.2502.14669}
 
 \leavevmode\vadjust pre{\hypertarget{ref-gh_Ehsan_2022}{}}%
 Ehsan, E. (2022). \emph{Maze}. \url{https://github.com/emadehsan/maze}
@@ -708,7 +711,7 @@ Ehsan, E. (2022). \emph{Maze}. \url{https://github.com/emadehsan/maze}
 Fung, S. W., Heaton, H., Li, Q., McKenzie, D., Osher, S., \& Yin, W.
 (2022). Jfb: Jacobian-free backpropagation for implicit networks.
 \emph{Proceedings of the AAAI Conference on Artificial Intelligence},
-\emph{36}, 6648--6656.
+\emph{36}, 6648--6656. \url{https://doi.org/10.1609/aaai.v36i6.20619}
 
 \leavevmode\vadjust pre{\hypertarget{ref-mathematica-maze}{}}%
 Guo, C., Barthelet, L., \& Morris, R. (2011). \emph{Maze generator and
@@ -719,11 +722,11 @@ solver}. Wolfram Demonstrations Project,
 Harries, L., Lee, S., Rzepecki, J., Hofmann, K., \& Devlin, S. (n.d.).
 {MazeExplorer}: {A Customisable 3D Benchmark} for {Assessing
 Generalisation} in {Reinforcement Learning}. \emph{2019 {IEEE Conf}.
-{Games CoG}}, 1--4.
+{Games CoG}}, 1--4. \url{https://doi.org/10.1109/cig.2019.8848048}
 
 \leavevmode\vadjust pre{\hypertarget{ref-zanj}{}}%
 Ivanitskiy, M. (n.d.). \emph{ZANJ}.
-\url{https://github.com/mivanit/ZANJ}
+\url{https://doi.org/10.5281/zenodo.15540393}
 
 \leavevmode\vadjust pre{\hypertarget{ref-maze-dataset-github}{}}%
 Ivanitskiy, Michael I., Shah, R., Spies, A. F., Räuker, T., Valentine,
@@ -742,34 +745,38 @@ Ivanitskiy, Michael Igorevich, Shah, R., Spies, A. F., Räuker, T.,
[…]
