+++ docs/paper/refs.bib
@@ -51,7 +51,8 @@
   volume={5},
   number={46},
   pages={1815},
-  year={2020}
+  year={2020},
+  doi = {10.21105/joss.01815},
 }
 
 @misc{mathematica-maze,
@@ -70,7 +71,8 @@
   number={1},
   pages={48--50},
   year={1956},
-  publisher={JSTOR}
+  publisher={JSTOR},
+  doi = {10.2307/2033241},
 }
 
 @article{pysr,
@@ -106,7 +108,8 @@
   volume={40},
   pages={353--359},
   year={2008},
-  publisher={Springer}
+  publisher={Springer},
+  doi = {10.3758/brm.40.1.353},
 }
 
 
@@ -118,7 +121,8 @@
   booktitle={2023 IEEE Conference on Secure and Trustworthy Machine Learning (SaTML)},
   pages={464--483},
   year={2023},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/satml54575.2023.00039},
 }
 
 @misc{percolation,
@@ -149,7 +153,8 @@
     author={Vaswani, Ashish and Shazeer, Noam and Parmar, Niki and Uszkoreit, Jakob and Jones, Llion and Gomez, Aidan N and Kaiser, {\L}ukasz and Polosukhin, Illia},
     journal={Advances in neural information processing systems},
     volume={30},
-    year={2017}
+    year={2017},
+    doi = {10.48550/arXiv.1706.03762},
 }
 
 @article{fukushima1983neocognitron,
@@ -159,7 +164,8 @@
     number={5},
     pages={826--834},
     year={1983},
-    publisher={IEEE}
+    publisher={IEEE},
+    doi = {10.1016/0893-6080(88)90014-7},
 }
 
 
@@ -224,6 +230,7 @@
     number = {6},
     pages = {1389--1401},
     publisher = {{Nokia Bell Labs}},
+    doi = {10.1002/j.1538-7305.1957.tb01515.x},
 }
 
 @misc{easy_to_hard,
@@ -275,6 +282,7 @@
     volume={34},
     pages={6695--6706},
     year={2021},
+    doi = {10.48550/arXiv.2106.04537},
 }
 
 @article{eval-gpt-visual,
@@ -317,7 +325,9 @@
     title = {{PyTorch: An Imperative Style, High-Performance Deep Learning Library}},
     url = {http://papers.neurips.cc/paper/9015-pytorch-an-imperative-style-high-performance-deep-learning-library.pdf},
     year = {2019},
+    doi = {10.48550/arXiv.1912.01703},
 }
+
 @misc{gh_Ehsan_2022, 
     title={Maze}, 
     url={https://github.com/emadehsan/maze}, 
@@ -337,6 +347,7 @@
     url={https://github.com/mivanit/ZANJ}, 
     abstractNote={saving complicated structures to disk without pickling}, 
     author={Ivanitskiy, Michael},
+    doi = {10.5281/zenodo.15540393},
 }
 
 @article{ivanitskiy2023structuredworldreps,
@@ -377,7 +388,8 @@
   journal={Advances in Neural Information Processing Systems},
   volume={37},
   pages={112329--112355},
-  year={2024}
+  year={2024},
+  doi = {10.48550/arXiv.2406.05183},
 }
 
 
@@ -417,11 +429,12 @@
 
 
 @inproceedings{harriesMazeExplorerCustomisable3D2019,
-    author = {Harries, Luke and Lee, Sebastian and Rzepecki, Jaroslaw and Hofmann, Katja and Devlin, Sam},
[…]
