+++ docs/paper/paper.tex
@@ -731,14 +731,13 @@ Ivanitskiy, M. (n.d.). \emph{ZANJ}.
 \leavevmode\vadjust pre{\hypertarget{ref-maze-dataset-github}{}}%
 Ivanitskiy, Michael I., Shah, R., Spies, A. F., Räuker, T., Valentine,
 D., Rager, C., Quirke, L., Corlouer, G., \& Mathwin, C. (2023a).
-\emph{Maze dataset}.
-\url{https://github.com/understanding-search/maze-dataset}
+\emph{Maze dataset}. \url{https://doi.org/10.48550/arXiv.2309.10498}
 
 \leavevmode\vadjust pre{\hypertarget{ref-maze-transformer-github}{}}%
 Ivanitskiy, Michael I., Shah, R., Spies, A. F., Räuker, T., Valentine,
 D., Rager, C., Quirke, L., Corlouer, G., \& Mathwin, C. (2023b).
 \emph{Maze transformer interpretability}.
-\url{https://github.com/understanding-search/maze-transformer}
+\url{https://doi.org/10.48550/arXiv.2312.02566}
 
 \leavevmode\vadjust pre{\hypertarget{ref-maze-dataset-arxiv-2023}{}}%
 Ivanitskiy, Michael Igorevich, Shah, R., Spies, A. F., Räuker, T.,
