+++ docs/paper/refs.bib
@@ -308,12 +308,14 @@
     url={https://github.com/understanding-search/maze-dataset},
     author={Ivanitskiy, Michael I. and Shah, Rusheb and Spies, Alex F. and Räuker, Tilman and Valentine, Dan and Rager, Can and Quirke, Lucia and Corlouer, Guillaume and Mathwin, Chris}, 
     year={2023},
+    doi = {10.48550/arXiv.2309.10498},
 }
 @misc{maze-transformer-github, 
     title={Maze Transformer Interpretability}, 
     url={https://github.com/understanding-search/maze-transformer},
     author={Ivanitskiy, Michael I. and Shah, Rusheb and Spies, Alex F. and Räuker, Tilman and Valentine, Dan and Rager, Can and Quirke, Lucia and Corlouer, Guillaume and Mathwin, Chris},
     year={2023},
+    doi = {10.48550/arXiv.2312.02566},
 }
 
 @inproceedings{pytorch,
