### 2025-05-29 #8042 @mivanit
Hello @ekourlit,

Thank you for taking over as editor!

We have reduced the paper to just over 1000 words as requested. We have done our best to resolve problems with DOIs, but we cite a number of repositories and websites which do not have DOIs (I have checked Zenodo for these).

I believe that the authors of the [Easy-to-Hard Generalization](https://arxiv.org/pdf/2403.09472) paper would be very qualified to review this work, as would the author or reviewers of [MDL Suite](https://joss.theoj.org/papers/10.21105/joss.01815), as our work is very related to those works. There would not be any conflict of interest with those individuals, to my knowledge.

Please let me know what further edits or actions we might need to take before we proceed with the review.
