+++ docs/paper/figures/outputs-ascii-colored.tex
@@ -8,7 +8,7 @@
 \# &   &   &   &   & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \# &  & \# \\
 \# &   & \# & \# & \# & \textcolor{blue}{X} & \# & \textcolor{blue}{X} & \# &   & \# \\
 \# &   &   &   & \# & \textcolor{blue}{X} & \# & \textcolor{green}{S} &   &   & \# \\
-\# & \# & \# & \# & \# & \textcolor{b			lue}{X} & \# & \# & \# & \# & \# \\
+\# & \# & \# & \# & \# & \textcolor{blue}{X} & \# & \# & \# & \# & \# \\
 \# & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \textcolor{blue}{X} & \# & \textcolor{red}{E} & \textcolor{blue}{X} & \textcolor{blue}{X} & \# \\
 \# & \textcolor{blue}{X} & \# & \# & \# &   & \# & \# & \# & \textcolor{blue}{X} & \# \\
 \# & \textcolor{blue}{X} & \# &   &   &   &   &   & \# & \textcolor{blue}{X} & \# \\
