### 2025-06-11 #8042 @ekourlit
@akashnag would you be able to review this submission? I'm contacting you directly here because you have previously published in the JOSS.
