### 2025-07-03 #8042 @afspies
Hi All,

Perhaps @https://github.com/Jemoka might make a suitable reviewer? It seems he has reviewed for JOSS before and has plenty of domain expertise.  The paper has started to accrue a fair few citations and it would be nice to refine the paper/package, and start collecting non-arvix citations :)

Thanks all!
