### 2025-07-26 #8633 @amcandio
Hi @mivanit! I noticed that the repository currently uses the Creative Commons Attribution-ShareAlike 4.0 International (CC BY-SA 4.0) license. I don't believe this license is OSI-approved and therefore does not meet the JOSS requirement for software. See https://opensource.org/license.

If possible, could you update the repository to include a plain-text LICENSE file with the contents of an OSI-approved software license?
