+++ docs/paper/preamble.tex
@@ -25,3 +25,5 @@
 \def\maxheight{\ifdim\Gin@nat@height>0.8\textheight 0.8\textheight\else\Gin@nat@height\fi}
 \setkeys{Gin}{width=\maxwidth,height=\maxheight,keepaspectratio}
 \makeatother
+\usepackage{textcomp}
+\newcommand{\mathparagraph}{\textparagraph}
