+++ docs/paper/preamble.tex
@@ -26,4 +26,4 @@
 \setkeys{Gin}{width=\maxwidth,height=\maxheight,keepaspectratio}
 \makeatother
 \usepackage{textcomp}
-\newcommand{\mathparagraph}{\textparagraph}
+\providecommand{\mathparagraph}{\textparagraph}
