+++ docs/other/lmcat.txt
@@ -1,8 +1,8 @@
 # Stats
 - 64 files
-- 16400 (16K) lines
-- 532427 (532K) chars
-- 60025 (60K) `whitespace-split` tokens
+- 16144 (16K) lines
+- 520579 (521K) chars
+- 58499 (58K) `whitespace-split` tokens
 
 # File Tree
 
@@ -80,8 +80,8 @@ maze-dataset
 │   ├── output_formats.ipynb               [  235L    20,199C 1,348T]
 │   └── profile_dataset_save_read.ipynb    [1,133L   269,709C 3,406T]
 ├── templates                              
-├── LICENSE.md                             [  427L    20,126C 2,768T]
-├── README.md                              [  186L     8,054C   696T]
+├── LICENSE.md                             [  165L     7,652C 1,234T]
+├── README.md                              [  191L     8,680C   704T]
 ├── makefile                               [1,851L    57,868C 6,652T]
 ├── pyproject.toml                         [  364L    12,919C 1,470T]
 ```
@@ -15062,433 +15062,172 @@ SPEEDS["load_minimal:profiling"][len(SPEEDS) - 1].sort_stats("tottime").print_st
 ``````{ end_of_file="notebooks/profile_dataset_save_read.ipynb" }
 
 ``````{ path="LICENSE.md"  }
-Attribution-ShareAlike 4.0 International
-
-=======================================================================
-
-Creative Commons Corporation ("Creative Commons") is not a law firm and
-does not provide legal services or legal advice. Distribution of
-Creative Commons public licenses does not create a lawyer-client or
-other relationship. Creative Commons makes its licenses and related
-information available on an "as-is" basis. Creative Commons gives no
-warranties regarding its licenses, any material licensed under their
-terms and conditions, or any related information. Creative Commons
-disclaims all liability for damages resulting from their use to the
-fullest extent possible.
-
-Using Creative Commons Public Licenses
-
-Creative Commons public licenses provide a standard set of terms and
-conditions that creators and other rights holders may use to share
-original works of authorship and other material subject to copyright
-and certain other rights specified in the public license below. The
-following considerations are for informational purposes only, are not
-exhaustive, and do not form part of our licenses.
-
-     Considerations for licensors: Our public licenses are
-     intended for use by those authorized to give the public
-     permission to use material in ways otherwise restricted by
-     copyright and certain other rights. Our licenses are
-     irrevocable. Licensors should read and understand the terms
-     and conditions of the license they choose before applying it.
-     Licensors should also secure all rights necessary before
-     applying our licenses so that the public can reuse the
-     material as expected. Licensors should clearly mark any
-     material not subject to the license. This includes other CC-
-     licensed material, or material used under an exception or
-     limitation to copyright. More considerations for licensors:
-	wiki.creativecommons.org/Considerations_for_licensors
-
-     Considerations for the public: By using one of our public
-     licenses, a licensor grants the public permission to use the
-     licensed material under specified terms and conditions. If
-     the licensor's permission is not necessary for any reason--for
-     example, because of any applicable exception or limitation to
-     copyright--then that use is not regulated by the license. Our
-     licenses grant only permissions under copyright and certain
-     other rights that a licensor has authority to grant. Use of
-     the licensed material may still be restricted for other
-     reasons, including because others have copyright or other
-     rights in the material. A licensor may make special requests,
-     such as asking that all changes be marked or described.
-     Although not required by our licenses, you are encouraged to
-     respect those requests where reasonable. More_considerations
-     for the public:
-	wiki.creativecommons.org/Considerations_for_licensees
-
-=======================================================================
-
-Creative Commons Attribution-ShareAlike 4.0 International Public
-License
-
[…]
