### 2025-07-27 26ea6d2
fix umlaut in paper?
