+++ docs/paper/paper.tex
@@ -15,7 +15,7 @@
     pdftitle={maze-dataset: Maze Generation with Algorithmic Variety and
 Representational Flexibility},
     pdfauthor={Michael Igorevich Ivanitskiy, Aaron Sandoval, Alex F.
-Spies, Tilman Räuker, Brandon Knutson, Cecilia Diniz Behn, Samy Wu
+Spies, Tilman R"auker, Brandon Knutson, Cecilia Diniz Behn, Samy Wu
 Fung},
     pdfpublication={Journal of Open Source Software},
     pdfpublisher={Open Journals},
@@ -25,7 +25,7 @@ Fung},
     pdfissuenum={},
     pdfdoi={N/A},
     pdfcopyright={Copyright (c) 1970, Michael Igorevich Ivanitskiy,
-Aaron Sandoval, Alex F. Spies, Tilman Räuker, Brandon Knutson, Cecilia
+Aaron Sandoval, Alex F. Spies, Tilman R"auker, Brandon Knutson, Cecilia
 Diniz Behn, Samy Wu Fung},
     pdflicenseurl={http://creativecommons.org/licenses/by/4.0/},
     colorlinks=true,
@@ -322,7 +322,7 @@ Representational Flexibility}
 }
 \author[3%
 %
-]{Tilman Räuker%
+]{Tilman R"auker%
   \,\orcidlink{0009-0009-6321-4413}\,%
 }
 \author[1%
