+++ docs/paper/refs.bib
@@ -115,7 +115,7 @@
 
 
 
-@inproceedings{interpretability-survery,
+@inproceedings{interpretability-survey,
   title={Toward transparent ai: A survey on interpreting the inner structures of deep neural networks},
   author={R{\"a}uker, Tilman and Ho, Anson and Casper, Stephen and Hadfield-Menell, Dylan},
   booktitle={2023 IEEE Conference on Secure and Trustworthy Machine Learning (SaTML)},
