+++ docs/paper/paper.tex
@@ -15,7 +15,7 @@
     pdftitle={maze-dataset: Maze Generation with Algorithmic Variety and
 Representational Flexibility},
     pdfauthor={Michael Igorevich Ivanitskiy, Aaron Sandoval, Alex F.
-Spies, Tilman R"auker, Brandon Knutson, Cecilia Diniz Behn, Samy Wu
+Spies, Tilman Räuker, Brandon Knutson, Cecilia Diniz Behn, Samy Wu
 Fung},
     pdfpublication={Journal of Open Source Software},
     pdfpublisher={Open Journals},
@@ -25,7 +25,7 @@ Fung},
     pdfissuenum={},
     pdfdoi={N/A},
     pdfcopyright={Copyright (c) 1970, Michael Igorevich Ivanitskiy,
-Aaron Sandoval, Alex F. Spies, Tilman R"auker, Brandon Knutson, Cecilia
+Aaron Sandoval, Alex F. Spies, Tilman Räuker, Brandon Knutson, Cecilia
 Diniz Behn, Samy Wu Fung},
     pdflicenseurl={http://creativecommons.org/licenses/by/4.0/},
     colorlinks=true,
@@ -322,7 +322,7 @@ Representational Flexibility}
 }
 \author[3%
 %
-]{Tilman R"auker%
+]{Tilman Räuker%
   \,\orcidlink{0009-0009-6321-4413}\,%
 }
 \author[1%
@@ -442,7 +442,7 @@ Language Models (LLMs) and other text-based transformer models.
 For interpretability and behavioral research, algorithmic tasks offer
 benefits by allowing systematic data generation and task decomposition,
 as well as simplifying the process of circuit discovery
-(\protect\hyperlink{ref-interpretability-survery}{Räuker et al., 2023}).
+(\protect\hyperlink{ref-interpretability-survey}{Räuker et al., 2023}).
 Although mazes are well suited for these investigations, we found that
 existing maze generation packages
 (\protect\hyperlink{ref-cobbe2019procgen}{Cobbe et al., 2019};
@@ -474,21 +474,21 @@ with these datasets.
   \begin{itemize}
   \tightlist
   \item
-    \href{https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/rasterized.html\#RasterizedMazeDataset}{\texttt{RasterizedMazeDataset}},
+    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/rasterized.html#RasterizedMazeDataset}{\texttt{RasterizedMazeDataset}}}{RasterizedMazeDataset},
     utilizing
-    \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_pixels}{\texttt{as\_pixels()}},
+    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html#LatticeMaze.as_pixels}{\texttt{as\_pixels()}}}{as\_pixels()},
     which can exactly mimic the outputs provided in
     \texttt{easy-to-hard-data}(\protect\hyperlink{ref-easy_to_hard}{Schwarzschild,
     Borgnia, Gupta, Bansal, et al., 2021}) and can be configured to be
     similar to the outputs of Németh
     (\protect\hyperlink{ref-gh_Nemeth_2019}{2019})
   \item
-    \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}
+    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}}{as\_ascii()}
[…]
