### 2025-07-27 e78699f
try again to fix LaTeX error
