+++ docs/paper/paper.tex
@@ -474,22 +474,22 @@ with these datasets.
   \begin{itemize}
   \tightlist
   \item
-    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/rasterized.html#RasterizedMazeDataset}{\texttt{RasterizedMazeDataset}}}{RasterizedMazeDataset},
+    \href{https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/rasterized.html\#RasterizedMazeDataset}{\texttt{RasterizedMazeDataset}},
     utilizing
-    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html#LatticeMaze.as_pixels}{\texttt{as\_pixels()}}}{as\_pixels()},
+    \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_pixels}{\texttt{as\_pixels()}},
     which can exactly mimic the outputs provided in
-    \texttt{easy-to-hard-data}(\protect\hyperlink{ref-easy_to_hard}{Schwarzschild,
-    Borgnia, Gupta, Bansal, et al., 2021}) and can be configured to be
-    similar to the outputs of Németh
-    (\protect\hyperlink{ref-gh_Nemeth_2019}{2019})
+    \texttt{easy-to-hard-data}
+    (\protect\hyperlink{ref-easy_to_hard}{Schwarzschild, Borgnia, Gupta,
+    Bansal, et al., 2021}) and can be configured to be similar to the
+    outputs of Németh (\protect\hyperlink{ref-gh_Nemeth_2019}{2019})
   \item
-    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}}{as\_ascii()}
+    \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}
     provides a format similar to
     (\protect\hyperlink{ref-gh-oppenheimj2018maze}{Oppenheim, 2018};
     \protect\hyperlink{ref-eval-gpt-visual}{Singla, 2023})
   \item
-    \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset/plotting.html#MazePlot}{\texttt{MazePlot}}}{MazePlot}
-    provides a feature-rich plotting utility with support for multiple
+    \href{https://understanding-search.github.io/maze-dataset/maze_dataset/plotting.html\#MazePlot}{\texttt{MazePlot}}
+    provides a feature‑rich plotting utility with support for multiple
     paths, heatmaps over positions, and more. This is similar to the
     outputs of (\protect\hyperlink{ref-mazegenerator-net}{Alance AB,
     2019}; \protect\hyperlink{ref-gh_Ehsan_2022}{Ehsan, 2022};
@@ -498,7 +498,7 @@ with these datasets.
   \end{itemize}
 \item
   The text format provided by
-  \texorpdfstring{\href{https://understanding-search.github.io/maze-dataset/maze_dataset.html#MazeDataset.as_tokens}{\texttt{SolvedMaze(...).as\_tokens()}}}{SolvedMaze(...).as\_tokens()}
+  \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#MazeDataset.as_tokens}{\texttt{SolvedMaze(...).as\_tokens()}}
   is similar to that of (\protect\hyperlink{ref-eval-LLM-graphs}{Liu \&
   Wu, 2023}) but with many more options, detailed
   \secref{sec:tokenized-output-formats}.
