### 2025-07-27 20a3b6c
yet another attempt to get latex in CI to work
