+++ docs/paper/preamble.tex
@@ -27,5 +27,5 @@
 \makeatother
 
 \pdfstringdefDisableCommands{%
-  \def\href#1#2{#2}%  % keep the visible text, drop the URL
+  \def\href##1##2{##2}%  % keep visible text, drop the URL
 }
