### 2025-07-28 41b3eed
paper-docker recipe for locally reproducing latex issue
