### 2025-07-28 understanding-search/maze-dataset#70 @amcandio
JOSS review: contributing and support guidelines
Hi! I see there is a development section in the README but the repository should include clear and more detailed community guidelines. Specifically:

- How can others contribute (e.g., CONTRIBUTING.md or a section in the README)?
- How should users report bugs or issues (e.g., via GitHub Issues)?
- Where can users seek support or ask questions?

Consider adding a short “Contributing” or “Community” section to the README, and/or linking to a `CONTRIBUTING.md` and using issue templates if available.

https://github.com/openjournals/joss-reviews/issues/8633
