### 2025-07-28 understanding-search/maze-dataset#71 @amcandio
JOSS review: citing dependencies
While reviewing dependencies I noted some of the dependencies are not cited in the paper. For example, numpy must be cited https://www.nature.com/articles/s41586-020-2649-2. Can you add citations for the relevant dependencies?

On a separate note, can you improve formating of pyproject.toml file?

https://github.com/openjournals/joss-reviews/issues/8633
