### 2025-07-28 understanding-search/maze-dataset#72 @amcandio
JOSS review: example in README.md raises exception
After doing `pip install maze-dataset` and then following examples in `README.md` I get:

```
Traceback (most recent call last):
  File "<python-input-5>", line 7, in <module>
    m.as_tokens(maze_tokenizer=MazeTokenizerModular(
                               ~~~~~~~~~~~~~~~~~~~~^
        tokenization_mode=TokenizationMode.AOTP_UT_rasterized, max_grid_size=100,
        ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
    ))
    ^
```
