### 2025-07-28 understanding-search/maze-dataset#73 @amcandio
JOSS review: installation issues
Instalation guide should clarify `uv` requirement and potential issues with Mac. Either having the `make dep` handle that or having `make dep_mac` should be enough to clarify the latter. You can also consider mentioning https://github.com/understanding-search/maze-dataset/issues/57 in the README.

### Context:
When installing dependencies I get the following error
```
(venv) bash-3.2$ make dep
Exporting dependencies as per pyproject.toml section 'tool.uv-exports.exports'
uv sync --all-extras --all-groups --compile-bytecode
make: uv: No such file or directory
make: *** [dep] Error 1
```

I was able to fix that by installing `uv`. Then I got this error:

```
(venv) bash-3.2$ make dep
...
Creating virtual environment at: .venv
Resolved 183 packages in 7ms
      Built maze-dataset @ file:///Users/candioti/workplace/maze-dataset
  × Failed to build `rust-fst==0.1.2`
  ├─▶ The build backend returned an error
  ╰─▶ Call to `setuptools.build_meta:__legacy__.build_wheel` failed (exit status: 1)

      [stderr]
      Traceback (most recent call last):
       ...
      ModuleNotFoundError: No module named 'rust_setuptools'

      hint: This usually indicates a problem with the package or the build environment.
  help: `rust-fst` (v0.1.2) was included because `maze-dataset[tokenization]` (v1.3.2) depends on `rust-fst`
make: *** [dep] Error 1
```

https://github.com/openjournals/joss-reviews/issues/8633

Which I see is related to https://github.com/understanding-search/maze-dataset/issues/57.
