### 2025-07-28 understanding-search/maze-dataset#74 @amcandio
JOSS review: testing issues
When running `make test` I'm facing the following issues:

### 1. Import errors:

Seems like `frozendict` and `pysr` were not included as part of project dependencies.
```
_____________________________________________________ ERROR collecting tests/unit/generation/test_solved_maze.py
import_module
tests/unit/generation/test_solved_maze.py:5: in <module>
    from maze_dataset.testing_utils import LEGACY_AND_EQUIVALENT_TOKENIZERS
maze_dataset/testing_utils.py:9: in <module>
    import frozendict
E   ModuleNotFoundError: No module named 'frozendict'
```

```
Traceback (most recent call last):
  File ".../maze-dataset/notebooks/../tests/_temp/notebooks/estimate_dataset_fractions.py", line 15, in <module>
    import pysr  # noqa: F401
    ^^^^^^^^^^^
ModuleNotFoundError: No module named 'pysr'
```

### 2. Notebook errors taking too long

Not an issue per se but would be good for users to be able to run those tests separatedly and have `README.md` file explain how to.

### 3. rust_fst import errors

All tests seem to pass but I'm seeing this error at the end which is related to https://github.com/understanding-search/maze-dataset/issues/57. Test with this dependency should be able to be skipped.

```
    from rust_fst import Set as FstSet  # type: ignore[import-untyped]
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
ModuleNotFoundError: No module named 'rust_fst'
```

After catching that import I see:

```
Traceback (most recent call last):
  File "<frozen runpy>", line 198, in _run_module_as_main
  File "<frozen runpy>", line 88, in _run_code
  File "maze-dataset/maze_dataset/tokenization/modular/fst.py", line 198, in <module>
    check_tokenizers_fst(
    ~~~~~~~~~~~~~~~~~~~~^
    	verbose=not args.quiet,
     ^^^^^^^^^^^^^^^^^^^^^^^
    	parallel=args.parallel,
     ^^^^^^^^^^^^^^^^^^^^^^^
    	n_check=n_check,
     ^^^^^^^^^^^^^^^^
    )
    ^
  File "maze-dataset/maze_dataset/tokenization/modular/fst.py", line 99, in check_tokenizers_fst
    get_tokenizers_fst()
    ~~~~~~~~~~~~~~~~~~^^
  File "maze-dataset/maze_dataset/tokenization/modular/fst_load.py", line 47, in get_tokenizers_fst
    return FstSet(MMT_FST_PATH.as_posix())
           ^^^^^^
NameError: name 'FstSet' is not defined
```

https://github.com/openjournals/joss-reviews/issues/8633
