### 2025-07-28 understanding-search/maze-dataset#75 @amcandio
JOSS review: dead links and images in doc site
In https://understanding-search.github.io/maze-dataset/maze_dataset.html I see some deadlinks and missing images.

https://github.com/openjournals/joss-reviews/issues/8633
