### 2025-07-28 understanding-search/maze-dataset#76 @Jemoka
`make dep` command fails for specific build
On some UV builds, the `--all-groups` flag doesn't exist, making dependency installation challenging. See error below:
```
$ make dep
Exporting dependencies as per pyproject.toml section 'tool.uv-exports.exports'
uv sync --all-extras --all-groups --compile-bytecode
error: unexpected argument '--all-groups' found

  tip: a similar argument exists: '--group'

Usage: uv sync --all-extras --group <GROUP>

For more information, try '--help'.
make: *** [dep] Error 2
```

Here are my version info:

```
$ uv --version
uv 0.4.28 (debe67ffd 2024-10-28)
```

on `macOS 14.2 (23C64)`, with M2 chip.

Created for https://github.com/openjournals/joss-reviews/issues/8633.
