### 2025-07-28 #8633 @Jemoka
Done with first round. Issues I experienced that are still unresolved

- https://github.com/understanding-search/maze-dataset/issues/70
- https://github.com/understanding-search/maze-dataset/issues/72
- https://github.com/understanding-search/maze-dataset/issues/74
- https://github.com/understanding-search/maze-dataset/issues/76

Looks like also the paper isn't building on JOSS's end. Should we trigger a rebuild? The software looks great, but the infrastructure / installation instructions could be made clearer, and the setup instruction crashes needs to be resolved.

Thanks for the great package!
