### 2025-07-28 understanding-search/maze-dataset#68 @mivanit
Thanks for pointing this out!

- @mivanit: core code contributor, initiated project and co-lead on AISC project
- @aaron-sandoval: core code contributor, and specifically led most of the tokenizer refactor
- @afspies: contributed code to this repo, also was one of the primary users of the repository and provided extremely valuable testing. also a co-lead on AISC project
- @bknutson0: contributed code, another primary user of the repository (for a different project), also provided very valuable testing
- @traeuker played a key supervisory role, and was also a co-lead on the AISC project. His contributions were critical for project direction
- @swufung and @cdinizbehn played key supervisory roles as well, as well as helping review and improve documentation

I should also point out that because this repository was originally part of https://github.com/understanding-search/maze-transformer before being spun out into its own library, the commit history of this repo underrepresents the contributions of @afspies and @traeuker. For clarity, we have verified with the relevant contributors to the previous project that they do not request authorship in this work, and they are listed in the acknowledgements.
