### 2025-07-28 1cf5770
clean up old comments in pyprojec.toml
