### 2025-07-28 understanding-search/maze-dataset#76 @mivanit
Unfortunately, I think trying to support `uv` builds that old is somewhat impractical. A lot of things have changed with the uv interface in the past year, which is now on `0.8.3`
