### 2025-07-28 understanding-search/maze-dataset#75 @mivanit
I actually also noticed this, and I think it should be fixed as of merging #69
