### 2025-07-28 891e718
explain uv in readme.md development section
