### 2025-07-28 understanding-search/maze-dataset#76 @mivanit
added a note about supported uv versions in [891e718](https://github.com/understanding-search/maze-dataset/pull/77/commits/891e7183fba5bfbc92d35f796a1cc2b322787516)
