### 2025-07-28 e2f5ba4
fix diagram width in readme/docs index
