### 2025-07-28 1be9a94
add 'contributing' section to readme.md
