### 2025-07-28 6d85b14
test and fix readme examples
