### 2025-07-28 58d814a
rework dependencies
