+++ .meta/requirements/requirements-all.txt
@@ -1,13 +1,13 @@
 # This file was autogenerated by uv via the following command:
-#    uv export --no-hashes --group dev --group lint --group benchmark --all-extras
+#    uv export --no-hashes --group dev --group lint --group docs --group benchmark --group extra-tokenization --all-extras
 -e .
-anyio==4.8.0
+anyio==4.9.0
     # via
     #   httpx
     #   jupyter-server
 appnope==0.1.4 ; sys_platform == 'darwin'
     # via ipykernel
-argon2-cffi==23.1.0
+argon2-cffi==25.1.0
     # via jupyter-server
 argon2-cffi-bindings==21.2.0
     # via argon2-cffi
@@ -15,9 +15,9 @@ arrow==1.3.0
     # via isoduration
 asttokens==3.0.0
     # via stack-data
-async-lru==2.0.4
+async-lru==2.0.5
     # via jupyterlab
-attrs==25.1.0
+attrs==25.3.0
     # via
     #   igittigitt
     #   jsonschema
@@ -26,13 +26,13 @@ babel==2.17.0
     # via jupyterlab-server
 backports-tarfile==1.2.0 ; python_full_version < '3.12' and platform_machine != 'ppc64le' and platform_machine != 's390x'
     # via jaraco-context
-beautifulsoup4==4.13.3
+beautifulsoup4==4.13.4
     # via nbconvert
 bleach==6.2.0
     # via nbconvert
-bracex==2.5.post1 ; python_full_version >= '3.11'
+bracex==2.6 ; python_full_version >= '3.11'
     # via wcmatch
-certifi==2025.1.31
+certifi==2025.7.14
     # via
     #   httpcore
     #   httpx
@@ -43,11 +43,11 @@ cffi==1.17.1
     #   cryptography
     #   pyzmq
     #   rust-fst
-charset-normalizer==3.4.1
+charset-normalizer==3.4.2
     # via requests
 cli-exit-tools==1.2.7 ; python_full_version >= '3.11'
     # via igittigitt
-click==8.1.8
+click==8.2.1
     # via
     #   cli-exit-tools
     #   igittigitt
@@ -58,22 +58,24 @@ colorama==0.4.6 ; sys_platform == 'win32'
     #   ipython
     #   pytest
     #   tqdm
-comm==0.2.2
+comm==0.2.3
     # via
     #   ipykernel
     #   ipywidgets
-contourpy==1.3.1
+contourpy==1.3.2 ; python_full_version < '3.11'
     # via matplotlib
-coverage==7.6.12
+contourpy==1.3.3 ; python_full_version >= '3.11'
+    # via matplotlib
+coverage==7.10.1
     # via
     #   coverage-badge
     #   pytest-cov
 coverage-badge==1.1.2
-cryptography==44.0.2 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
+cryptography==45.0.5 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
     # via secretstorage
 cycler==0.12.1
     # via matplotlib
-debugpy==1.8.13
+debugpy==1.8.15
     # via ipykernel
 decorator==5.2.1
     # via ipython
@@ -81,7 +83,7 @@ defusedxml==0.7.1
     # via nbconvert
 docutils==0.21.2
     # via readme-renderer
-exceptiongroup==1.2.2 ; python_full_version < '3.11'
+exceptiongroup==1.3.0 ; python_full_version < '3.11'
     # via
     #   anyio
     #   ipython
[…]
