+++ .meta/requirements/requirements-user.txt
@@ -1,5 +1,5 @@
 # This file was autogenerated by uv via the following command:
-#    uv export --no-hashes --group dev --group lint --group docs --group benchmark --group extra-tokenization --extra tokenization
+#    uv export --no-hashes --no-group dev --no-group lint --no-group docs --no-group benchmark --no-group extra-tokenization --extra tokenization
 -e .
 anyio==4.9.0
     # via
@@ -19,19 +19,14 @@ async-lru==2.0.5
     # via jupyterlab
 attrs==25.3.0
     # via
-    #   igittigitt
     #   jsonschema
     #   referencing
 babel==2.17.0
     # via jupyterlab-server
-backports-tarfile==1.2.0 ; python_full_version < '3.12' and platform_machine != 'ppc64le' and platform_machine != 's390x'
-    # via jaraco-context
 beautifulsoup4==4.13.4
     # via nbconvert
 bleach==6.2.0
     # via nbconvert
-bracex==2.6 ; python_full_version >= '3.11'
-    # via wcmatch
 certifi==2025.7.14
     # via
     #   httpcore
@@ -40,23 +35,13 @@ certifi==2025.7.14
 cffi==1.17.1
     # via
     #   argon2-cffi-bindings
-    #   cryptography
     #   pyzmq
     #   rust-fst
 charset-normalizer==3.4.2
     # via requests
-cli-exit-tools==1.2.7 ; python_full_version >= '3.11'
-    # via igittigitt
-click==8.2.1
-    # via
-    #   cli-exit-tools
-    #   igittigitt
-    #   pysr
 colorama==0.4.6 ; sys_platform == 'win32'
     # via
-    #   click
     #   ipython
-    #   pytest
     #   tqdm
 comm==0.2.3
     # via
@@ -66,13 +51,6 @@ contourpy==1.3.2 ; python_full_version < '3.11'
     # via matplotlib
 contourpy==1.3.3 ; python_full_version >= '3.11'
     # via matplotlib
-coverage==7.10.1
-    # via
-    #   coverage-badge
-    #   pytest-cov
-coverage-badge==1.1.2
-cryptography==45.0.5 ; platform_machine != 'ppc64le' and platform_machine != 's390x' and sys_platform == 'linux'
-    # via secretstorage
 cycler==0.12.1
     # via matplotlib
 debugpy==1.8.15
@@ -81,21 +59,14 @@ decorator==5.2.1
     # via ipython
 defusedxml==0.7.1
     # via nbconvert
-docutils==0.21.2
-    # via readme-renderer
 exceptiongroup==1.3.0 ; python_full_version < '3.11'
     # via
     #   anyio
     #   ipython
-    #   pytest
-execnet==2.1.1
-    # via pytest-xdist
 executing==2.2.0
     # via stack-data
 fastjsonschema==2.21.1
     # via nbformat
-filelock==3.18.0
-    # via juliapkg
 fonttools==4.59.0
     # via matplotlib
 fqdn==1.5.1
@@ -108,27 +79,18 @@ httpcore==1.0.9
     # via httpx
 httpx==0.28.1
     # via jupyterlab
-id==1.5.0
-    # via twine
 idna==3.10
     # via
     #   anyio
     #   httpx
     #   jsonschema
     #   requests
-igittigitt==2.1.5 ; python_full_version >= '3.11'
-    # via lmcat
-importlib-metadata==8.7.0 ; python_full_version < '3.12' and platform_machine != 'ppc64le' and platform_machine != 's390x'
-    # via keyring
-iniconfig==2.1.0
-    # via pytest
[…]
