### 2025-07-28 2dd8762
add warning for when on macos
