### 2025-07-28 88431be
add warning about macOS to readme
