### 2025-07-28 7c175be
fix more ruff issues
