### 2025-07-28 understanding-search/maze-dataset#73 @mivanit
In #77 I've added a note about uv being required to the readme, thanks for pointing that out!

Can you clarify what OS you're on? The relevant dep is marked as
```
rust_fst>=0.1.2; platform_system != 'darwin'
```
which means it should not be installed on macOS -- however, I'm unable to test this as I don't have a macOS system.
