### 2025-07-29 understanding-search/maze-dataset#68 @amcandio
Thanks for the clarification, marking as resolved.
