### 2025-07-29 understanding-search/maze-dataset#73 @amcandio
@mivanit even with the marker I'm seeing the issue. I tried replacing `'darwin'` by `'Darwin'` and that worked!
