### 2025-07-29 understanding-search/maze-dataset#71 @amcandio
> Can you clarify what you mean by improving the formatting of the pyproject.toml file?

Oh sorry for the ambiguity. I just meant indent seems odd in the file. For example, these lines https://github.com/understanding-search/maze-dataset/blob/main/pyproject.toml#L52-L59
