+++ .meta/requirements/requirements-user.txt
@@ -314,7 +314,7 @@ rpds-py==0.26.0
     # via
     #   jsonschema
     #   referencing
-rust-fst==0.1.2 ; platform_system != 'darwin'
+rust-fst==0.1.2 ; sys_platform != 'darwin'
     # via maze-dataset
 send2trash==1.8.3
     # via jupyter-server
