### 2025-07-30 understanding-search/maze-dataset#71 @mivanit
> > Can you clarify what you mean by improving the formatting of the pyproject.toml file?
>
> Oh sorry for the ambiguity. I just meant that indent seems odd in the file. For example, these lines https://github.com/understanding-search/maze-dataset/blob/main/pyproject.toml#L52-L59

I do agree that this is not exactly standard, but it's compliant with the TOML spec, self-consistent, and in my opinion makes the lengthy file somewhat easier to read.
