### 2025-07-30 d57a3ee
better explanation of test recipes in README.md
see point 2 in https://github.com/understanding-search/maze-dataset/issues/74#issue-3267903156
