### 2025-07-30 2250b56
dont do pysr stuff during notebook tests
it's really, really slow. longer than all the other notebook tests combined, since it involves installing julia
