### 2025-07-30 b9f15b3
type checking
