+++ pyproject.toml
@@ -146,6 +146,8 @@
 		"tests/unit/processing/test_get_forking_path_points.py", # 58
 		# extra low priority (test temp, generated from notebooks)
 		"tests/_temp/*",
+		# don't check docs
+		"docs/*",
     ]
     check_untyped_defs = true
 
