### 2025-07-30 72d6812
add info on adding generators to docs
see https://github.com/understanding-search/maze-dataset/issues/78
