### 2025-07-30 understanding-search/maze-dataset#74 @mivanit
this should now be resolved in #77
1. those have been added to dev dependencies
2. see commit [d57a3ee](https://github.com/understanding-search/maze-dataset/pull/77/commits/d57a3eef52225473394a32b3673993fefa0e2b6a)
3. tests using rust_fst will now just warn instead of error. I'm not sure if this is the correct behavior, might need to add a flag for this later. #57 will stay open to deal with this at some point, probably via migration to another FST library
