### 2025-07-30 98a2959
add citations to packages used
https://github.com/understanding-search/maze-dataset/issues/71
