### 2025-07-31 understanding-search/maze-dataset#70 @mivanit
should be fixed in #77 (section in README.md, not separate file), but some feedback on whether this is sufficient would definitely be helpful! I've never written a contributing guide before.

Also relates to:
- #78 since that part of the documentation is linked to in the same section of the README
- #72
- #73 since we update installation instructions in the README
- #75 since that also modifies the README
