### 2025-07-31 understanding-search/maze-dataset#72 @amcandio
Now I'm getting this error after having done `make dep`:

```
Exporting dependencies as per pyproject.toml section 'tool.uv-exports.exports'
warning: `VIRTUAL_ENV=venv` does not match the project environment path `.venv` and will be ignored; use `--active` to target the active environment instead
<string>:3: UserWarning:
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
WARNING: some tokenization features do not work on macOS, see https://github.com/understanding-search/maze-dataset/issues/57
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
uv sync --all-extras --all-groups --compile-bytecode
warning: `VIRTUAL_ENV=venv` does not match the project environment path `.venv` and will be ignored; use `--active` to target the active environment instead
Resolved 189 packages in 2ms
Bytecode compiled 9917 files in 1.59s
warning: `VIRTUAL_ENV=venv` does not match the project environment path `.venv` and will be ignored; use `--active` to target the active environment instead
+ uv export --no-hashes --no-group dev --no-group lint --no-group docs --no-group benchmark --no-group extra-tokenization
Resolved 189 packages in 1ms
+ uv export --no-hashes --group dev --group lint --group docs --group benchmark --group extra-tokenization --extra tokenization
Resolved 189 packages in 1ms
+ uv export --no-hashes --no-group dev --no-group lint --no-group docs --no-group benchmark --no-group extra-tokenization --extra tokenization
Resolved 189 packages in 1ms
+ uv export --no-hashes --only-group lint
Resolved 189 packages in 1ms
+ uv export --no-hashes --group dev --group lint --group docs --group benchmark --group extra-tokenization --extra tokenization
Resolved 189 packages in 2ms
+ uv export --no-hashes --group dev --group lint --group docs --group benchmark --group extra-tokenization --all-extras
Resolved 189 packages in 2ms
```

```
In [1]: from maze_dataset import MazeDataset, MazeDatasetConfig
   ...: from maze_dataset.generation import LatticeMazeGenerators
   ...: cfg: MazeDatasetConfig = MazeDatasetConfig(
   ...: ^Iname="test", # name is only for you to keep track of things
   ...: ^Igrid_n=5, # number of rows/columns in the lattice
   ...: ^In_mazes=4, # number of mazes to generate
   ...: ^Imaze_ctor=LatticeMazeGenerators.gen_dfs, # algorithm to generate the maze
   ...:     maze_ctor_kwargs=dict(do_forks=False), # additional parameters to pass to the maze generation algorithm
   ...: )
Cell In[1], line 1
----> 1 from maze_dataset import MazeDataset, MazeDatasetConfig
      2 from maze_dataset.generation import LatticeMazeGenerators
      3 cfg: MazeDatasetConfig = MazeDatasetConfig(
      4 	name="test", # name is only for you to keep track of things
      5 	grid_n=5, # number of rows/columns in the lattice
   (...)      8     maze_ctor_kwargs=dict(do_forks=False), # additional parameters to pass to the maze generation algorithm
      9 )

File ~/workplace/maze-dataset/maze_dataset/__init__.py:3
      1 """.. include:: ../README.md"""
----> 3 from maze_dataset.constants import (
      4 	SPECIAL_TOKENS,
      5 	VOCAB,
      6 	VOCAB_LIST,
      7 	VOCAB_TOKEN_TO_INDEX,
      8 	Connection,
      9 	ConnectionArray,
     10 	ConnectionList,
     11 	Coord,
     12 	CoordArray,
     13 	CoordList,
     14 	CoordTup,
     15 )
     16 from maze_dataset.dataset.collected_dataset import (
     17 	MazeDatasetCollection,
     18 	MazeDatasetCollectionConfig,
     19 )
     20 from maze_dataset.dataset.filters import register_maze_filter

File ~/workplace/maze-dataset/maze_dataset/constants.py:7
      4 from dataclasses import dataclass, field, make_dataclass
      5 from typing import Iterator
----> 7 import numpy as np
      8 from jaxtyping import Bool, Int8
     10 from maze_dataset.utils import corner_first_ndindex
```
