### 2025-07-31 understanding-search/maze-dataset#73 @amcandio
Works now, thanks!
