### 2025-07-31 understanding-search/maze-dataset#74 @amcandio
I'm still seeing this error after running the tests. You can probably reproduce this by uninstalling rust_fst and running `make test`

```
ModuleNotFoundError: No module named 'rust_fst'
make: *** [tokenizer-fst-check-small] Error 1
```
