### 2025-07-31 understanding-search/maze-dataset#74 @mivanit
Can you clarify which stage of the tests you are seeing the failure on?
`make test` is defined as:
```
.PHONY: test
test: clean test-unit test-notebooks-muutils tokenizer-fst-check-small
```
When uninstalling `rust_fst`, `test-unit` and `test-notebook-muutils` both pass, while `tokenizer-fst-check-small` fails.

I think that this is the correct behavior here -- unit tests and notebook tests pass, but the tokenizer fst checks fail. The latter tests are already separate, and getting them to pass without `rust_fst` would amount to the tests not doing anything.

I do, however, think that I should more clearly explain this behavior in the readme and documentation. Perhaps adding to the `# Development` section something like:

> Note: due to compatibility issues with the `rust_fst` package and Darwin/macOS systems, not all tests will pass on these systems. However, `make test-unit` and `make test-notebooks-muutils` should still pass. Please see #57 for updates on resolving this problem.
