### 2025-08-03 understanding-search/maze-dataset#74 @mivanit
added this in [9d60a03](https://github.com/understanding-search/maze-dataset/pull/77/commits/9d60a031656ac02970bc3f15fa510b4a1a7298f0), should be resolved now!
