### 2025-08-03 understanding-search/maze-dataset#74 @amcandio
Nice, all issues resolved here so closing!
