### 2025-08-04 429a2b1
rename numpy rng global
