### 2025-08-04 be3bad1
add a way to register new generator functions
