+++ tests/unit/generation/test_custom_generator.py
@@ -1,0 +1,299 @@
+"""Tests for custom maze generator registration system"""
+
+from pathlib import Path
+
+import numpy as np
+import pytest
+from zanj import ZANJ
+
+from maze_dataset import LatticeMaze, MazeDataset, MazeDatasetConfig
+from maze_dataset.constants import Coord, CoordTup
+from maze_dataset.generation import (
+	GENERATORS_MAP,
+	LatticeMazeGenerators,
+	get_maze_with_solution,
+)
+from maze_dataset.generation.registration import (
+	MazeGeneratorRegistrationError,
+	register_maze_generator,
+)
+
+# Temp directory for file operations
+TEMP_PATH = Path("tests/_temp/test_custom_generator/")
+
+
+def test_register_valid_function():
+	"""Test that a valid function can be registered successfully"""
+
+	@register_maze_generator
+	def gen_test_valid(
+		grid_shape: Coord | CoordTup,
+		lattice_dim: int = 2,
+	) -> LatticeMaze:
+		"""Simple test generator - fully connected grid"""
+		grid_shape_: Coord = np.array(grid_shape)
+		connection_list: np.ndarray = np.zeros(
+			(lattice_dim, *grid_shape_), dtype=np.bool_
+		)
+
+		# Create fully connected grid
+		if grid_shape_[1] > 1:
+			connection_list[1, :, : grid_shape_[1] - 1] = True
+		if grid_shape_[0] > 1:
+			connection_list[0, : grid_shape_[0] - 1, :] = True
+
+		return LatticeMaze(
+			connection_list=connection_list,
+			generation_meta=dict(
+				func_name="gen_test_valid",
+				grid_shape=grid_shape_,
+				fully_connected=True,
+			),
+		)
+
+	# Test registration worked
+	assert "gen_test_valid" in GENERATORS_MAP
+	assert hasattr(LatticeMazeGenerators, "gen_test_valid")
+
+	# Test function works
+	maze = get_maze_with_solution("gen_test_valid", (5, 5))
+	assert maze.maze.grid_shape == (5, 5)
+
+	# Test via LatticeMazeGenerators
+	maze2 = LatticeMazeGenerators.gen_test_valid((4, 4))
+	assert maze2.grid_shape == (4, 4)
+
+
+def test_maze_dataset_config_with_custom_generator():
+	"""Test creating, saving, and loading MazeDatasetConfig with custom generator"""
+
+	@register_maze_generator
+	def gen_test_config(
+		grid_shape: Coord | CoordTup,
+		custom_param: float = 0.5,
+	) -> LatticeMaze:
+		"""Test generator with custom parameter"""
+		grid_shape_: Coord = np.array(grid_shape)
+		connection_list: np.ndarray = np.zeros((2, *grid_shape_), dtype=np.bool_)
+
+		# Simple fully connected pattern
+		if grid_shape_[1] > 1:
+			connection_list[1, :, : grid_shape_[1] - 1] = True
+		if grid_shape_[0] > 1:
+			connection_list[0, : grid_shape_[0] - 1, :] = True
+
+		return LatticeMaze(
+			connection_list=connection_list,
+			generation_meta=dict(
+				func_name="gen_test_config",
+				grid_shape=grid_shape_,
+				custom_param=custom_param,
+				fully_connected=True,
+			),
+		)
+
+	# Create config with custom generator
+	config = MazeDatasetConfig(
+		name="test_custom",
+		grid_n=5,
+		n_mazes=3,
+		maze_ctor=gen_test_config,
+		maze_ctor_kwargs={"custom_param": 0.7},
+	)
+
+	# Test serialization/deserialization
+	serialized = config.serialize()
+	loaded_config = MazeDatasetConfig.load(serialized)
+
[…]
