+++ maze_dataset/generation/registration.py
@@ -1,0 +1,173 @@
+"""Registration system for custom maze generators"""
+
+import inspect
+from typing import Callable, Union, get_args, get_origin
+
+from maze_dataset.constants import Coord, CoordTup
+from maze_dataset.generation.generators import GENERATORS_MAP, LatticeMazeGenerators
+from maze_dataset.maze import LatticeMaze
+
+# Type for maze generator functions
+MazeGeneratorFunc = Callable[[Union[Coord, CoordTup], ...], LatticeMaze]
+
+
+class MazeGeneratorRegistrationError(TypeError):
+	"""error for maze generator registration issues"""
+
+	pass
+
+
+def _check_grid_shape_annotation(
+	annotation: Union[type, str],
+	func_name: str,
+) -> None:
+	"""Check if the annotation for grid_shape is valid
+
+	# Parameters:
+	- `annotation` : Union[type, str] the type annotation of the grid_shape parameter
+
+	# Raises:
+	- `MazeGeneratorRegistrationError` if the annotation is not compatible
+	"""
+	annotation_str: str = str(annotation)
+
+	if not any(
+		[
+			annotation == Coord,
+			annotation == CoordTup,
+			annotation_str.lower() == "tuple[int, int]",
+			# TODO: these are pretty loose checks, would be better to make this more robust
+			"ndarray" in annotation_str,
+			"Int8" in annotation_str,
+		]
+	):
+		err_msg = (
+			f"Maze generator function '{func_name}' first parameter 'grid_shape' "
+			f"must be typed as 'Coord | CoordTup' or compatible type, "
+			f"got {annotation = }, {annotation_str = }"
+		)
+		raise MazeGeneratorRegistrationError(err_msg)
+
+
+def validate_MazeGeneratorFunc(
+	func: MazeGeneratorFunc,
+) -> None:
+	"""validate the signature of a maze generator function
+
+	return `None` if valid, otherwise raises `MazeGeneratorRegistrationError`
+	(which is a subclass of `TypeError`)
+
+	# Parameters:
+	- `func : MazeGeneratorFunc` function to validate
+
+	# Raises:
+	- `MazeGeneratorRegistrationError` : type error describing the issue with the function signature
+	"""
+	func_name: str = func.__name__
+	sig: inspect.Signature = inspect.signature(func)
+	params: list[str] = list(sig.parameters.keys())
+
+	if not params or params[0] != "grid_shape":
+		err_msg = (
+			f"Maze generator function '{func_name}' must have 'grid_shape' "
+			"as its first parameter. Please ensure the function signature starts with 'grid_shape: Coord | CoordTup'."
+			f"{params = }"
+		)
+		raise MazeGeneratorRegistrationError(err_msg)
+
+	# Check first parameter type annotation if present
+	first_param_annotation = sig.parameters["grid_shape"].annotation
+	if first_param_annotation == inspect.Parameter.empty:
+		err_msg = (
+			f"Maze generator function '{func_name}' must have a type annotation for 'grid_shape'. "
+			"Please add `grid_shape: Coord | CoordTup` to the function signature."
+		)
+
+	# Check if it's a Union type
+	if get_origin(first_param_annotation) in (Union, type(Union[int, str])):
+		args: tuple = get_args(first_param_annotation)
+		# Check all of the args look like Coord or CoordTup
+		for arg in args:
+			_check_grid_shape_annotation(arg, func_name)
+	else:
+		# Check if the annotation is a single type
+		_check_grid_shape_annotation(first_param_annotation, func_name)
+
+	# Check return type annotation - must be present and correct
+	if sig.return_annotation == inspect.Signature.empty:
+		err_msg = (
+			f"Maze generator function '{func_name}' must have a return type annotation "
[…]
