+++ maze_dataset/generation/__init__.py
@@ -3,28 +3,28 @@
 `DEFAULT_GENERATORS` is a list of generator name, generator kwargs pairs used in tests and demos
 
 you can add your own maze generators by:
-- adding a static method implementing your generation function to `LatticeMazeGenerators`, with the signature `(grid_shape: Coord, **kwargs) -> LatticeMaze`
-- adding the pair to `GENERATORS_MAP`
+- adding a static method implementing your generation function to `LatticeMazeGenerators`, with the signature `(grid_shape: Coord, **kwargs) -> LatticeMaze` and adding the `(name, func)` pair to `GENERATORS_MAP`
+- using the `@register_maze_generator` decorator on your generation function. However, this is only for testing purposes where modifying the original package is not possible.
 
 If you implement a new generation function, please make a pull request!
 https://github.com/understanding-search/maze-dataset/pulls
 """
 
 from maze_dataset.generation.generators import (
+	_NUMPY_RNG,
 	GENERATORS_MAP,
 	LatticeMazeGenerators,
 	get_maze_with_solution,
-	numpy_rng,
 )
 
 __all__ = [
 	# submodules
 	"default_generators",
 	"generators",
-	"seed",
+	"registrationseed",
 	# imports
 	"LatticeMazeGenerators",
 	"GENERATORS_MAP",
 	"get_maze_with_solution",
-	"numpy_rng",
+	"_NUMPY_RNG",
 ]
