### 2025-08-04 46f7fac
regen docs, regen examples
