+++ docs/other/lmcat.txt
@@ -1,8 +1,8 @@
 # Stats
-- 64 files
-- 16137 (16K) lines
-- 520582 (521K) chars
-- 58506 (59K) `whitespace-split` tokens
+- 66 files
+- 16637 (17K) lines
+- 533593 (534K) chars
+- 60014 (60K) `whitespace-split` tokens
 
 # File Tree
 
@@ -10,11 +10,11 @@
 maze-dataset                               
 ├── .github                                
 │   └── workflows                          
-│    ├── all-tok-checks.yml                [   55L     1,728C   155T]
+│    ├── all-tok-checks.yml                [   53L     1,684C   152T]
 │    ├── build-paper.yml                   [   28L       785C    85T]
-│    ├── checks.yml                        [   94L     2,164C   224T]
-│    ├── make-docs.yml                     [   47L     1,262C   141T]
-│    └── speed-benchmark.yml               [   48L     1,261C   109T]
+│    ├── checks.yml                        [   90L     2,099C   217T]
+│    ├── make-docs.yml                     [   44L     1,198C   134T]
+│    └── speed-benchmark.yml               [   45L     1,191C   102T]
 ├── maze_dataset                           
 │   ├── benchmark                          
 │   │   ├── __init__.py                    [    9L       220C    24T]
@@ -25,20 +25,21 @@ maze-dataset
 │   │   ├── __init__.py                    [   45L     2,111C   243T]
 │   │   ├── collected_dataset.py           [  241L     8,113C   848T]
 │   │   ├── configs.py                     [  285L     8,000C   668T]
-│   │   ├── dataset.py                     [  553L    18,539C 2,133T]
-│   │   ├── filters.py                     [  295L     9,571C 1,100T]
-│   │   ├── maze_dataset.py                [  632L    20,310C 2,073T]
+│   │   ├── dataset.py                     [  548L    18,385C 2,123T]
+│   │   ├── filters.py                     [  295L     9,541C 1,094T]
+│   │   ├── maze_dataset.py                [  633L    20,369C 2,089T]
 │   │   ├── maze_dataset_config.py         [  496L    17,917C 2,019T]
-│   │   ├── rasterized.py                  [  330L     9,476C 1,001T]
+│   │   ├── rasterized.py                  [  330L     9,478C 1,001T]
 │   │   └── success_predict_math.py        [  106L     2,809C   460T]
 │   ├── generation                         
-│   │   ├── __init__.py                    [   23L       534C    54T]
+│   │   ├── __init__.py                    [   30L     1,109C   126T]
 │   │   ├── default_generators.py          [   23L       817C    72T]
-│   │   ├── generators.py                  [  651L    23,337C 2,932T]
+│   │   ├── generators.py                  [  660L    23,594C 2,961T]
+│   │   ├── registration.py                [  176L     5,731C   678T]
 │   │   └── seed.py                        [    3L        45C     7T]
 │   ├── maze                               
 │   │   ├── __init__.py                    [   34L     1,905C   272T]
-│   │   └── lattice_maze.py                [1,532L    49,541C 5,753T]
+│   │   └── lattice_maze.py                [1,542L    49,787C 5,787T]
 │   ├── plotting                           
 │   │   ├── __init__.py                    [   33L     1,008C   110T]
 │   │   ├── plot_dataset.py                [   64L     1,635C   200T]
[…]
