+++ docs/other/todo-inline-standard.md
@@ -98,9 +98,9 @@
 
 
 - idk why type ignore here  
-  local link: [`/maze_dataset/maze/lattice_maze.py:1421`](/maze_dataset/maze/lattice_maze.py#L1421) 
-  | view on GitHub: [maze_dataset/maze/lattice_maze.py#L1421](https://github.com/understanding-search/maze-dataset/blob/main/maze_dataset/maze/lattice_maze.py#L1421)
-  | [Make Issue](https://github.com/understanding-search/maze-dataset/issues/new?title=idk%20why%20type%20ignore%20here&body=%23%20source%0A%0A%5B%60maze_dataset%2Fmaze%2Flattice_maze.py%23L1421%60%5D%28https%3A%2F%2Fgithub.com%2Funderstanding-search%2Fmaze-dataset%2Fblob%2Fmain%2Fmaze_dataset%2Fmaze%2Flattice_maze.py%23L1421%29%0A%0A%23%20context%0A%60%60%60python%0A%09%09%22%22%22%0A%09%09forks_idxs%2C%20_%20%3D%20self.get_solution_forking_points%28%29%0A%09%09%23%20HACK%3A%20idk%20why%20type%20ignore%20here%0A%09%09return%20%28%20%20%23%20type%3A%20ignore%5Breturn-value%5D%0A%09%09%09np.delete%28np.arange%28self.solution.shape%5B0%5D%29%2C%20forks_idxs%2C%20axis%3D0%29%2C%0A%60%60%60&labels=HACK)
+  local link: [`/maze_dataset/maze/lattice_maze.py:1431`](/maze_dataset/maze/lattice_maze.py#L1431) 
+  | view on GitHub: [maze_dataset/maze/lattice_maze.py#L1431](https://github.com/understanding-search/maze-dataset/blob/main/maze_dataset/maze/lattice_maze.py#L1431)
+  | [Make Issue](https://github.com/understanding-search/maze-dataset/issues/new?title=idk%20why%20type%20ignore%20here&body=%23%20source%0A%0A%5B%60maze_dataset%2Fmaze%2Flattice_maze.py%23L1431%60%5D%28https%3A%2F%2Fgithub.com%2Funderstanding-search%2Fmaze-dataset%2Fblob%2Fmain%2Fmaze_dataset%2Fmaze%2Flattice_maze.py%23L1431%29%0A%0A%23%20context%0A%60%60%60python%0A%09%09%22%22%22%0A%09%09forks_idxs%2C%20_%20%3D%20self.get_solution_forking_points%28%29%0A%09%09%23%20HACK%3A%20idk%20why%20type%20ignore%20here%0A%09%09return%20%28%20%20%23%20type%3A%20ignore%5Breturn-value%5D%0A%09%09%09np.delete%28np.arange%28self.solution.shape%5B0%5D%29%2C%20forks_idxs%2C%20axis%3D0%29%2C%0A%60%60%60&labels=HACK)
 
   ```python
   """
@@ -218,9 +218,9 @@
 ## [`maze_dataset/dataset/dataset.py`](/maze_dataset/dataset/dataset.py)
 
 - get rid of all these things as part of migration to tokenizer-free dataset config  
-  local link: [`/maze_dataset/dataset/dataset.py:70`](/maze_dataset/dataset/dataset.py#L70) 
-  | view on GitHub: [maze_dataset/dataset/dataset.py#L70](https://github.com/understanding-search/maze-dataset/blob/main/maze_dataset/dataset/dataset.py#L70)
[…]
