+++ docs/coverage/coverage.txt
@@ -9,18 +9,19 @@ maze_dataset/constants.py                                        85     71    16
 maze_dataset/dataset/__init__.py                                  4      4     0%   23-30
 maze_dataset/dataset/collected_dataset.py                        87     49    44%   8-31, 42, 52-53, 57-58, 62-63, 67-68, 72, 76, 83-86, 106-107, 111-112, 116-117, 125, 129, 143-144, 156-169, 180-181, 192, 217, 227
 maze_dataset/dataset/configs.py                                  47     25    47%   85-285
-maze_dataset/dataset/dataset.py                                 201    121    40%   9-33, 39-45, 59-85, 90, 94, 100, 113-223, 257, 271-272, 278, 290, 294-315, 326, 333-334, 342-385, 389, 401-402, 406, 416-429, 444, 461-507, 521-526
+maze_dataset/dataset/dataset.py                                 202    122    40%   9-34, 40-46, 60-86, 91, 95, 101, 114-224, 258, 272-273, 279, 291, 295-316, 327, 334-335, 341-380, 384, 396-397, 401, 411-424, 439, 456-502, 516-521
 maze_dataset/dataset/filters.py                                 111     61    45%   3-20, 49-55, 59-61, 67-69, 87-102, 121, 151-152, 165-167, 178-191, 221, 232, 239-241, 255-262, 268-284
-maze_dataset/dataset/maze_dataset.py                            178     99    44%   7-44, 72, 80, 95-115, 128-129, 169, 173, 177, 181, 189-203, 226, 230, 234, 237, 243, 249-250, 325-326, 330-331, 336, 341, 344-350, 358-359, 379-428, 437, 448, 454, 490-540, 548, 568, 589
+maze_dataset/dataset/maze_dataset.py                            174     95    45%   7-44, 72, 80, 95-116, 129-130, 170, 174, 178, 182, 204, 227, 231, 235, 238, 244, 250-251, 326-327, 331-332, 337, 342, 345-351, 359-360, 380-429, 438, 449, 455, 491-541, 549, 569, 590
 maze_dataset/dataset/maze_dataset_config.py                     118     90    24%   3-55, 61, 66-113, 115, 131-140, 194-195, 199-200, 204-209, 229, 236, 245, 256-282, 286-287, 294, 304, 325-496
 maze_dataset/dataset/rasterized.py                               92     21    77%   233, 251-273, 303
 maze_dataset/dataset/success_predict_math.py                     20     20     0%   6-101
-maze_dataset/generation/__init__.py                               2      2     0%   6-13
+maze_dataset/generation/__init__.py                               2      2     0%   13-20
 maze_dataset/generation/default_generators.py                     1      0   100%
-maze_dataset/generation/generators.py                           206     35    83%   3-19, 37, 53-61, 101-105, 207-208, 230-231, 319-320, 362-363, 404-405, 513-514, 613-639
+maze_dataset/generation/generators.py                           208     37    82%   3-19, 37, 53-61, 102-106, 208-209, 231-232, 320-321, 363-364, 405-406, 514-515, 613-648
+maze_dataset/generation/registration.py                          43      3    93%   91-94
 maze_dataset/generation/seed.py                                   1      1     0%   3
 maze_dataset/maze/__init__.py                                     2      2     0%   11-22
[…]

[…]
