### 2025-08-04 68f1423
full coverage re-run
`make test-cov; make cov`
