### 2025-08-04 3a73898
narrowed it down to figs 4, 6, 7, 8 and tab 1 are the issue in paper compilation
