+++ docs/paper/paper.tex
@@ -23,7 +23,7 @@ Fung},
     pdfpubtype={journal},
     pdfvolumenum={},
     pdfissuenum={},
-    pdfdoi={N/A},
+    pdfdoi={10.xxxxxx/draft},
     pdfcopyright={Copyright (c) 1970, Michael Igorevich Ivanitskiy,
 Aaron Sandoval, Alex F. Spies, Tilman Räuker, Brandon Knutson, Cecilia
 Diniz Behn, Samy Wu Fung},
@@ -139,9 +139,10 @@ Diniz Behn, Samy Wu Fung},
 \renewcommand{\footrulewidth}{0.25pt}
 
 \fancyfoot[L]{\parbox[t]{0.98\headwidth}{\footnotesize{\sffamily Ivanitskiy
-et al. (1970). maze-dataset: Maze Generation with Algorithmic Variety
+et al. (2025). maze-dataset: Maze Generation with Algorithmic Variety
 and Representational Flexibility. \emph{Journal of Open Source
-Software}, \emph{¿VOL?}(¿ISSUE?), ¿PAGE? \url{https://doi.org/N/A}}.}}
+Software}, \emph{¿VOL?}(¿ISSUE?), ¿PAGE?
+\url{https://doi.org/10.xxxxxx/draft}}.}}
 
 
 \fancyfoot[R]{\sffamily \thepage}
@@ -277,6 +278,9 @@ Software}, \emph{¿VOL?}(¿ISSUE?), ¿PAGE? \url{https://doi.org/N/A}}.}}
 % get rid of language-specific shorthands (see #6817):
 \let\LanguageShortHands\languageshorthands
 \def\languageshorthands#1{}
+\usepackage{lineno}
+\linenumbers
+\usepackage{draftwatermark}
 
 \IfFileExists{parskip.sty}{%
 \usepackage{parskip}
@@ -362,7 +366,7 @@ and Statistics%
   %\hrule
   \sffamily\small
 
-  {\bfseries DOI:} \href{https://doi.org/N/A}{\color{linky}{N/A}}
+  {\bfseries DOI:} \href{https://doi.org/10.xxxxxx/draft}{\color{linky}{10.xxxxxx/draft}}
 
   \vspace{2mm}
     {\bfseries Software}
@@ -391,7 +395,7 @@ Journals} \ExternalLink
     \vspace{2mm}
   
     {\bfseries Submitted:} 01 January 1970\\
-    {\bfseries Published:} 01 January 1970
+    {\bfseries Published:} unpublished
 
   \vspace{2mm}
   {\bfseries License}\\
@@ -402,8 +406,7 @@ Journals} \ExternalLink
   \end{flushleft}
 }
 
-\hypertarget{summary}{%
-\section{Summary}\label{summary}}
+\section{Summary}\label{summary}
 
 Solving mazes is a classic problem in computer science and artificial
 intelligence, and humans have been constructing mazes for thousands of
@@ -425,8 +428,7 @@ is extensively tested, type hinted, benchmarked, and documented.
 
 \input{figures/tex/fig1_diagram.tex}
 
-\hypertarget{statement-of-need}{%
-\section{Statement of Need}\label{statement-of-need}}
+\section{Statement of Need}\label{statement-of-need}
 
 While maze generation itself is straightforward, the architectural
 challenge comes from building a system supporting many algorithms with
@@ -442,29 +444,27 @@ Language Models (LLMs) and other text-based transformer models.
 For interpretability and behavioral research, algorithmic tasks offer
 benefits by allowing systematic data generation and task decomposition,
 as well as simplifying the process of circuit discovery
-(\protect\hyperlink{ref-interpretability-survey}{Räuker et al., 2023}).
[…]
