+++ docs/paper/paper.tex
@@ -596,6 +596,8 @@ certain constraints may not always be successful, and for this we
 provide a way to estimate the success rate of a given configuration,
 described in \autoref{fig:sre}.
 
+\input{figures/tex/fig6_benchmarks.tex}
+
 \section{Implementation}\label{sec:implementation}
 
 Using an adjacency matrix for storing mazes would be memory inefficient
