+++ docs/paper/paper.tex
@@ -585,6 +585,8 @@ of text output formats, with an example given in
 (\citeproc{ref-Gallant2015Transducers}{Gallant, 2015}) for efficiently
 storing valid tokenizers.
 
+\input{figures/tex/fig4_tokenfmt.tex}
+
 \subsection{Benchmarks}\label{benchmarks}
 
 We benchmarks for generation time across various configurations in
