+++ docs/paper/paper.tex
@@ -598,6 +598,8 @@ certain constraints may not always be successful, and for this we
 provide a way to estimate the success rate of a given configuration,
 described in \autoref{fig:sre}.
 
+\input{figures/tex/tab1_benchmarks.tex}
+
 \input{figures/tex/fig6_benchmarks.tex}
 
 \section{Implementation}\label{sec:implementation}
