+++ docs/paper/paper.md
@@ -114,7 +114,7 @@ We benchmarks for generation time across various configurations in \autoref{tab:
 
 
 
-<!-- \input{figures/tex/tab1_benchmarks.tex} -->
+\input{figures/tex/tab1_benchmarks.tex}
 
 \input{figures/tex/fig6_benchmarks.tex}
 
