### 2025-08-04 a485ddb
fix figure 7
