+++ docs/paper/paper.tex
@@ -602,6 +602,8 @@ described in \autoref{fig:sre}.
 
 \input{figures/tex/fig6_benchmarks.tex}
 
+\input{figures/tex/fig7_sre.tex}
+
 \section{Implementation}\label{sec:implementation}
 
 Using an adjacency matrix for storing mazes would be memory inefficient
