+++ docs/paper/paper.md
@@ -118,7 +118,7 @@ We benchmarks for generation time across various configurations in \autoref{tab:
 
 \input{figures/tex/fig6_benchmarks.tex}
 
-<!-- \input{figures/tex/fig7_sre.tex} -->
+\input{figures/tex/fig7_sre.tex}
 
 # Implementation {#sec:implementation}
 
