### 2025-08-04 2fe322c
fix citations, full paper compiling but fig heights messed up
