+++ docs/paper/paper.tex
@@ -613,6 +613,8 @@ simple, efficient representation of mazes as subgraphs of a finite
 lattice, detailed in \autoref{fig:maze-impl}, which we call a
 \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze}{\texttt{LatticeMaze}}.
 
+\input{figures/tex/fig8_impl.tex}
+
 Our package is implemented in Python(\citeproc{ref-python}{Rossum,
 1995}), and makes use of the extensive scientific computing ecosystem,
 including NumPy (\citeproc{ref-numpy}{Harris et al., 2020}) for array
