### 2025-08-04 2692606
fix some figure sizing, all except diagram
