+++ docs/paper/figures/tex/fig7_sre.tex
@@ -1,6 +1,6 @@
 \begin{figure}
 	\centering
-	\includegraphics[width=1\textwidth,height=\textheight]{figures/ep/ep_deadends_unique-crop.pdf}
+	\includegraphics[width=1\textwidth]{figures/ep/ep_deadends_unique-crop.pdf}
 	\caption{
 		In order to replicate the exact dataset distribution of (\citeproc{ref-easy_to_hard}{Schwarzschild, Borgnia, Gupta, Bansal, et
 		al., 2021}), % custom cite here due to weird interactions with inara
