### 2025-08-04 5f5860b
figure sizing, formatting edits
