+++ docs/paper/paper.tex
@@ -415,16 +415,15 @@ problem, this very fact makes it an excellent testbed for studying how
 machine learning algorithms solve problems and represent spatial
 information. We introduce \texttt{maze-dataset}, a user-friendly Python
 library for generating, processing, and visualizing datasets of mazes.
-This library supports a variety of maze generation algorithms providing
-mazes with or without loops, mazes that are connected or not, and many
-other variations. These generation algorithms can be configured with
-various parameters, and the resulting mazes can be filtered to satisfy
-desired properties. Also provided are tools for converting mazes to and
-from various formats suitable for a variety of neural network
-architectures, such as rasterized images, tokenized text sequences, and
-various visualizations. As well as providing a simple interface for
-generating, storing, and loading these datasets, \texttt{maze-dataset}
-is extensively tested, type hinted, benchmarked, and documented.
+This library supports a variety of maze generation algorithms which can
+be configured with various parameters, and the resulting mazes can be
+filtered to satisfy desired properties. Also provided are tools for
+converting mazes to and from various formats suitable for a variety of
+neural network architectures, such as rasterized images, tokenized text
+sequences, and various visualizations. As well as providing a simple
+interface for generating, storing, and loading these datasets,
+\texttt{maze-dataset} is extensively tested, type hinted, benchmarked,
+and documented.
 
 \input{figures/tex/fig1_diagram.tex}
 
