+++ docs/paper/refs.bib
@@ -412,6 +412,8 @@
 }
 
 
+% papers that use maze-dataset
+
 @article{chen2024iaimperative,
   title={{iA*}: Imperative Learning-based {A*} Search for Pathfinding},
   author={Chen, Xiangyu and Yang, Fan and Wang, Chen},
@@ -428,6 +430,43 @@
   doi = {10.48550/arXiv.2410.16162},
 }
 
+
+@article{cai2025morse,
+  title={MORSE-500: A Programmatically Controllable Video Benchmark to Stress-Test Multimodal Reasoning},
+  author={Cai, Zikui and Wang, Andrew and Satheesh, Anirudh and Nakhawa, Ankit and Jae, Hyunwoo and Powell, Keenan and Liu, Minghui and Jay, Neel and Oh, Sungbin and Wang, Xiyao and others},
+  journal={arXiv preprint arXiv:2506.05523},
+  year={2025},
+  doi = {10.48550/arXiv.2506.05523},
+}
+
+@article{xu2025visual,
+  title={Visual Planning: Let's Think Only with Images},
+  author={Xu, Yi and Li, Chengzu and Zhou, Han and Wan, Xingchen and Zhang, Caiqi and Korhonen, Anna and Vuli{\'c}, Ivan},
+  journal={arXiv preprint arXiv:2505.11409},
+  year={2025},
+  doi = {10.48550/arXiv.2505.11409},
+}
+
+@article{lee2025adaptive,
+  title={Adaptive Cyclic Diffusion for Inference Scaling},
+  author={Lee, Gyubin and Bao, Truong Nhat Nguyen and Yoon, Jaesik and Lee, Dongwoo and Kim, Minsu and Bengio, Yoshua and Ahn, Sungjin},
+  journal={arXiv preprint arXiv:2505.14036},
+  year={2025},
+  doi = {10.48550/arXiv.2505.14036},
+}
+
+
+@article{zhang2025vfscale,
+  title={VFScale: Intrinsic Reasoning through Verifier-Free Test-time Scalable Diffusion Model},
+  author={Zhang, Tao and Pan, Jia-Shu and Feng, Ruiqi and Wu, Tailin},
+  journal={arXiv preprint arXiv:2502.01989},
+  year={2025},
+  doi = {10.48550/arXiv.2502.01989},
+}
+
+
+
+% other?
 
 
 @inproceedings{harriesMazeExplorerCustomisable3D2019,
