+++ docs/paper/figures/tex/fig6_benchmarks.tex
@@ -1,7 +1,7 @@
 \begin{figure}
 	\hypertarget{fig:benchmarks}{%
 		\centering
-		\includegraphics[width=0.9\textwidth]{figures/benchmarks/gridsize-vs-gentime.pdf}
+		\includegraphics[width=\textwidth]{figures/benchmarks/gridsize-vs-gentime.pdf}
 		\caption{
 			Plot of maze generation time. Generation time scales
 			exponentially with maze size for all algorithms. Generation time per
