### 2025-08-04 #8633 @mivanit
- [maze-dataset#77](https://github.com/understanding-search/maze-dataset/pull/77) merged
- pdf compiles
- have **not** yet pushed to pypi - going to check that things don't break for some downstream users, expect to push the new version (1.4.0) in the next day

current pypi version:
