### 2025-08-13 #8633 @mivanit
I should clarify -- the only thing that doesn't work as a result of [#57](https://github.com/understanding-search/maze-dataset/issues/57) is the *validation* of a created tokenizer being in the tested set. Users can still create and use custom tokenizers.
