### 2025-08-21 bcdbec5
add .PHONY for recipe
