+++ docs/paper/figures/tex/fig1_diagram.tex
@@ -3,7 +3,7 @@
     	\input{diagram/diagram.tikz}
 	}
 	\caption{
-	  Usage of maze-dataset. We create a \texttt{MazeDataset} from a \texttt{MazeDatasetConfig}. This contains \texttt{SolvedMaze} objects which can be converted to and from a variety of formats. Code in the image contains clickable links to \docslink{maze_dataset.html}{documentation}. A variety of generated examples can be viewed \docslink{examples/maze_examples.html}{here}.
+	  Usage of maze-dataset. We create a \texttt{MazeDataset} from a \texttt{MazeDatasetConfig}. This contains \texttt{SolvedMaze} objects which can be converted to and from a variety of formats. A variety of generated examples can be viewed on the \docslink{examples/maze_examples.html}{examples page}, and more information can be found in the \docslink{maze_dataset.html}{documentation}.
 	}
 	\label{fig:diagram}
 \end{figure}
