+++ docs/paper/figures/tex/fig6_benchmarks.tex
@@ -4,7 +4,7 @@
 		\includegraphics[width=\textwidth]{figures/benchmarks/gridsize-vs-gentime.pdf}
 		\caption{
 			Plot of maze generation time. Generation time scales
-			exponentially with maze size for all algorithms. Generation time per
+			roughly exponentially with maze size for all algorithms. Generation time per
 			maze does not depend on the number of mazes being generated, and there
 			is minimal overhead to initializing the generation process for a small
 			dataset. Wilson's algorithm is notably less efficient than others and
