### 2025-08-21 5c66776
reorder sections, implementation before benchmarks
