+++ docs/paper/figures/tex/fig8_impl.tex
@@ -1,21 +1,21 @@
 \begin{figure}
 \centering
 \begin{tabular}{ccc}
-	Maze 
+	Maze \hspace{1em} 
 	& $A[0]$ (downward)
 	& $A[1]$ (rightward) \\
 	\raisebox{-0.5\height}{
-		\includegraphics[width=5em]{figures/tex/maze-impl.pdf}
-	}
+		\includegraphics[width=7em]{figures/tex/maze-impl.pdf}
+	} \hspace{1em}
 	& \raisebox{-0.5em}{
-		\large $\begin{bmatrix}
+		\huge $\begin{bmatrix}
 			1 & 1 & 1 \\
 			0 & 0 & 1 \\
 			0 & 0 & 0
 		\end{bmatrix}$
 	}
 	& \raisebox{-0.5em}{
-		\large $\begin{bmatrix}
+		\huge $\begin{bmatrix}
 			1 & 0 & 0 \\
 			0 & 1 & 0 \\
 			1 & 1 & 0
