### 2025-08-21 #8633 @mivanit
Regarding the post-review author tasks @ekourlit:
- [x] went over the paper again, and made some minor formatting edits
  - latest paper version compiles, see latest editorialbot message: https://raw.githubusercontent.com/openjournals/joss-papers/joss.08633/joss.08633/10.21105.joss.08633.pdf
- [x] released version 1.4.0 with all the changes from the review:
  - github: https://github.com/understanding-search/maze-dataset/releases/tag/v1.4.0
  - pypi: https://pypi.org/project/maze-dataset/1.4.0/
- [x] we have archived the release on Zenodo:
  - DOI: `10.5281/zenodo.16915901`
  - zenodo page: https://zenodo.org/records/16915901
- [x] double checked all author affiliations and ORCIDs, and made sure that the zenodo ORCIDs match
- [x] made sure both the zenodo posting and github repo have the LGPL v3 license
