### 2025-09-09 #8633 @ekourlit
Hi @mivanit , I had a final look at the paper and here are some minor proposed fixed before we press the acceptance button.
1. Synch authors names with ORCIDs (e.g. Alex F. Spies -> Alexander F. Spies).
2. In caption of Figure 2: styleze "numpy" to "NumPy", capitalize "feature-rich".
3. Fix the legend of Figure 7, so it is either inside the plot or completely out of it, without being cut by the render.
