### 2025-10-12 2b432ae
JOSS paper fixes
see https://github.com/openjournals/joss-reviews/issues/8633#issuecomment-3270514168

- `numpy` -> NumPy in figure 2
- proper capitalization of "feature-rich" in figure 2
- fixed legend in figure 7
- author name consistency: "Alex F. Spies" -> "Alexander F. Spies"
