+++ docs/paper/paper.tex
@@ -14,8 +14,8 @@
     pdfaconformance=B,
     pdftitle={maze-dataset: Maze Generation with Algorithmic Variety and
 Representational Flexibility},
-    pdfauthor={Michael Igorevich Ivanitskiy, Aaron Sandoval, Alex F.
-Spies, Tilman Räuker, Brandon Knutson, Cecilia Diniz Behn, Samy Wu
+    pdfauthor={Michael Igorevich Ivanitskiy, Aaron Sandoval, Alexander
+F. Spies, Tilman Räuker, Brandon Knutson, Cecilia Diniz Behn, Samy Wu
 Fung},
     pdfpublication={Journal of Open Source Software},
     pdfpublisher={Open Journals},
@@ -25,8 +25,8 @@ Fung},
     pdfissuenum={},
     pdfdoi={10.xxxxxx/draft},
     pdfcopyright={Copyright (c) 1970, Michael Igorevich Ivanitskiy,
-Aaron Sandoval, Alex F. Spies, Tilman Räuker, Brandon Knutson, Cecilia
-Diniz Behn, Samy Wu Fung},
+Aaron Sandoval, Alexander F. Spies, Tilman Räuker, Brandon Knutson,
+Cecilia Diniz Behn, Samy Wu Fung},
     pdflicenseurl={http://creativecommons.org/licenses/by/4.0/},
     colorlinks=true,
     linkcolor=[rgb]{0.0, 0.5, 1.0},
@@ -321,7 +321,7 @@ Representational Flexibility}
 }
 \author[2%
 %
-]{Alex F. Spies%
+]{Alexander F. Spies%
   \,\orcidlink{0000-0002-8708-1530}\,%
 }
 \author[3%
@@ -489,7 +489,7 @@ these datasets.
     \citeproc{ref-eval-gpt-visual}{Singla, 2023})
   \item
     \href{https://understanding-search.github.io/maze-dataset/maze_dataset/plotting.html\#MazePlot}{\texttt{MazePlot}}
-    provides a feature‑rich plotting utility with support for multiple
+    provides a feature-rich plotting utility with support for multiple
     paths, heatmaps over positions, and more. This is similar to the
     outputs of (\citeproc{ref-mazegenerator-net}{Alance AB, 2019};
     \citeproc{ref-gh_Ehsan_2022}{Ehsan, 2022};
