+++ docs/paper/figures/tex/fig2_formats.tex
@@ -8,8 +8,8 @@
 	  & \docslink{maze_dataset/plotting.html\#MazePlot}{\texttt{MazePlot()}} \\[.5em]
 	  % descriptions
 		Simple text format for displaying mazes, useful for debugging in a terminal environment.
-		& \texttt{numpy} array of \texttt{dtype=uint8} and shape \texttt{(height, width, 3)}. The last dimension is RGB color.
-		& feature-rich plotting utility with support for multiple paths, heatmaps over positions, and more. \\[1em]
+		& NumPy array of \texttt{dtype=uint8} and shape \texttt{(height, width, 3)}. The last dimension is RGB color.
+		& Feature-rich plotting utility with support for multiple paths, heatmaps over positions, and more. \\[1em]
 	  \hline \\
 	  % examples
 		\multicolumn{1}{l}{
