### 2025-10-12 269f74a
[joss paper] minor figure 2 edits
