### 2025-10-17 #8633 @crvernon
@openjournals/dev do you know why the following error is occurring?  https://github.com/openjournals/joss-reviews/issues/8633#issuecomment-3415740107
