+++ docs/paper/figures/tex/fig3_raster.tex
@@ -1,10 +1,8 @@
 \begin{figure}
-	\hypertarget{fig:e2h-raster}{%
 	\centering
 	\includegraphics[width=0.3\textwidth]{figures/maze-raster-input-target.pdf}
-	\caption{Input is the rasterized maze without the path marked (left),
-	and provide as a target the maze with all but the correct path removed
-	(right). Configuration options exist to adjust whether endpoints are
-	included and if empty cells should be filled in.}\label{fig:e2h-raster}
+	\caption{
+		Input is the rasterized maze without the path marked (left), and provide as a target the maze with all but the correct path removed (right). Configuration options exist to adjust whether endpoints are included and if empty cells should be filled in.
 	}
+	\label{fig:e2h-raster}
 \end{figure}
