### 2025-10-20 #8633 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
