### 2025-10-20 eeb06c5
paranthetical to narrative citations
or fix grammar for paranthetical ones to work, as in "This is similar to the outputs of..."

see https://github.com/openjournals/joss-reviews/issues/8633#issuecomment-3421967248
