### 2025-10-20 1dd5506
fix spacing in PySR ref
see https://github.com/openjournals/joss-reviews/issues/8633#issuecomment-3421967248
