+++ docs/paper/refs.bib
@@ -76,7 +76,7 @@
 }
 
 @article{pysr,
-  title={Interpretable machine learning for science with PySR and SymbolicRegression. jl},
+  title={Interpretable machine learning for science with PySR and SymbolicRegression.jl},
   author={Cranmer, Miles},
   journal={arXiv preprint arXiv:2305.01582},
   year={2023},
