### 2025-10-20 25d4e4f
recompile paper
