+++ docs/paper/paper.tex
@@ -484,23 +484,23 @@ these datasets.
     Németh (\citeproc{ref-gh_Nemeth_2019}{2019})
   \item
     \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#LatticeMaze.as_ascii}{\texttt{as\_ascii()}}
-    provides a format similar to
-    (\citeproc{ref-gh-oppenheimj2018maze}{Oppenheim, 2018};
-    \citeproc{ref-eval-gpt-visual}{Singla, 2023})
+    provides a format similar to Singla
+    (\citeproc{ref-eval-gpt-visual}{2023}) and Oppenheim
+    (\citeproc{ref-gh-oppenheimj2018maze}{2018})
   \item
     \href{https://understanding-search.github.io/maze-dataset/maze_dataset/plotting.html\#MazePlot}{\texttt{MazePlot}}
     provides a feature-rich plotting utility with support for multiple
     paths, heatmaps over positions, and more. This is similar to the
-    outputs of (\citeproc{ref-mazegenerator-net}{Alance AB, 2019};
-    \citeproc{ref-gh_Ehsan_2022}{Ehsan, 2022};
+    outputs of many packages (\citeproc{ref-mazegenerator-net}{Alance
+    AB, 2019}; \citeproc{ref-gh_Ehsan_2022}{Ehsan, 2022};
     \citeproc{ref-mathematica-maze}{Guo et al., 2011};
-    \citeproc{ref-mdl-suite}{Nag, 2020})
+    \citeproc{ref-mdl-suite}{Nag, 2020}).
   \end{itemize}
 \item
   The text format provided by
   \href{https://understanding-search.github.io/maze-dataset/maze_dataset.html\#MazeDataset.as_tokens}{\texttt{SolvedMaze(...).as\_tokens()}}
-  is similar to that of (\citeproc{ref-eval-LLM-graphs}{Liu \& Wu,
-  2023}) but with many more options, detailed in
+  is similar to that of Liu \& Wu (\citeproc{ref-eval-LLM-graphs}{2023})
+  but with many more options, detailed in
   \hyperref[sec:tokenized-output-formats]{section: \textit{\nameref{sec:tokenized-output-formats}}}.
 \item
   Preserving metadata about the generation algorithm with the dataset
@@ -559,8 +559,8 @@ In previous work, maze tasks have been used with Recurrent Convolutional
 Neural Network (RCNN) derived architectures
 (\citeproc{ref-deepthinking}{Schwarzschild, Borgnia, Gupta, Huang, et
 al., 2021}). To facilitate the use of our package in this context, we
-replicate the format of (\citeproc{ref-easy_to_hard}{Schwarzschild,
-Borgnia, Gupta, Bansal, et al., 2021}) and provide the
+replicate the format of Schwarzschild, Borgnia, Gupta, Bansal, et al.
+(\citeproc{ref-easy_to_hard}{2021}) and provide the
 \href{https://understanding-search.github.io/maze-dataset/maze_dataset/dataset/rasterized.html\#RasterizedMazeDataset}{\texttt{RasterizedMazeDataset}}
 class which returns rasterized pairs of (input, target) mazes as shown
 in \autoref{fig:e2h-raster}.
@@ -598,7 +598,7 @@ lattice, detailed in \autoref{fig:maze-impl}, which we call a
 
 \input{figures/tex/fig8_impl.tex}
 
-Our package is implemented in Python(\citeproc{ref-python}{Rossum,
+Our package is implemented in Python (\citeproc{ref-python}{Rossum,
 1995}), and makes use of the extensive scientific computing ecosystem,
[…]
